Module conrod::backend

Traits and functionality related to Conrod's generic backend.

These modules describe an interface allowing users to use their own graphics, events and character caching systems with Conrod.

Note: Conrod currently heavily depends upon the piston graphics and event crates for enabling genericity over custom user backends. This dependency may change in the near future in favour of simplified conrod-specific backend trait.

Reexports

pub use self::event::{RawEvent, ToRawEvent};

Modules

event
The backend for conrod's event system.

graphics
Conrod's generic graphics backend.

Traits

Backend
A trait to be implemented by all backends to conrod.

CharacterCache
Stores characters in a buffer and loads them by demand.

Graphics
Implemented by all graphics back-ends.