conrod::backend::event::RawEvent

type RawEvent = Event<Input>;

The event type that is used by conrod to track inputs from the world. This can be thought of as the event type which is supplied by the window backend to drive the state of the Ui forward.

This type is primarily used within the Ui::handle_event method which immediately converts given user events to RawEvents via the ToRawEvent trait bound. The RawEvents are interpreted to create higher level UiEvents (such as DoubleClick, WidgetCapturesKeyboard, etc) which are stored for later processing by Widgets, which will occur during the call to Ui::set_widgets.