Struct conrod::backend::gfx::Renderer

pub struct Renderer<'a, R: Resources> { /* fields omitted */ }

Methods

impl<'a, R: Resources> Renderer<'a, R>

pub fn new<F: Factory<R>>(
factory: &mut F,
rtv: &RenderTargetView<R, ColorFormat>,
dpi_factor: f64
) -> Result<Self, RendererCreationError>

Important traits for Commands<'a>
impl<'a> Iterator for Commands<'a> type Item = Command<'a>;

pub fn commands(&self) -> Commands

Produce an Iterator yielding Commands.

pub fn fill<P, C>(
&mut self,
encoder: &mut Encoder<R, C>,
dims: (f32, f32),
primitives: P,
image_map: &Map<(ShaderResourceView<R, [f32; 4]>, (u32, u32))>
) where
P: PrimitiveWalker,
C: CommandBuffer<R>,

Fill the inner vertex and command buffers by translating the given primitives.

pub fn draw<F, C>(
&self,
factory: &mut F,
encoder: &mut Encoder<R, C>,
image_map: &Map<(ShaderResourceView<R, [f32; 4]>, (u32, u32))>
) where
F: Factory<R>,
C: CommandBuffer<R>,

Draws using the inner list of Commands to the given display.

Note: If you require more granular control over rendering, you may want to use the fill and commands methods separately. This method is simply a convenience wrapper around those methods for the case that the user does not require accessing or modifying conrod's draw parameters, uniforms or generated draw commands.