Module conrod::backend::glium

A glium backend for rendering conrod primitives.

Modules

glium
Easy-to-use, high-level, OpenGL3+ wrapper.

Structs

Commands
An iterator yielding Commands, produced by the Renderer::commands method.

GlyphCache
A rusttype GlyphCache along with a glium::texture::Texture2d for caching text on the GPU.

Renderer
A type used for translating render::Primitives into Commands that indicate how to draw the conrod GUI using glium.

Vertex
The Vertex type passed to the vertex shader.

Enums

Command
A Command describing a step in the drawing process.

Draw
A Command for drawing to the target.

DrawError
Possible errors that may occur during a call to Renderer::draw.

RendererCreationError
Possible errors that may occur during a call to Renderer::new.

Constants

FRAGMENT_SHADER_120
The fragment shader used within the glium::Program for OpenGL.

FRAGMENT_SHADER_140
The fragment shader used within the glium::Program for OpenGL.

FRAGMENT_SHADER_300_ES
The fragment shader used within the glium::Program for OpenGL ES.

MODE_GEOMETRY
Ignore tex and draw simple, colored 2D geometry.

MODE_IMAGE
Draw an image from the texture at tex in the fragment shader.

MODE_TEXT
Draw text from the text cache texture tex in the fragment shader.

VERTEX_SHADER_120
The vertex shader used within the glium::Program for OpenGL.

VERTEX_SHADER_140
The vertex shader used within the glium::Program for OpenGL.

VERTEX_SHADER_300_ES
The vertex shader used within the glium::Program for OpenGL ES.

Traits

TextureDimensions
Glium textures that have two dimensions.

Functions

draw_parameters
Default glium DrawParameters with alpha blending enabled.

gamma_srgb_to_linear
Converts gamma (brightness) from sRGB to linear color space.

program
Construct the glium shader program that can be used to render Vertexes.

text_texture_client_format
Return the optimal client format for the text texture given the version.

text_texture_uncompressed_float_format
Return the optimal uncompressed float format for the text texture given the version.