Struct conrod::backend::glium::glium::backend::glutin::glutin::ContextBuilder

pub struct ContextBuilder<'a> {
    pub gl_attr: GlAttributes<&'a Context>,
    // some fields omitted
}

Object that allows you to build Contexts.

Fields
gl_attr: GlAttributes<&'a Context>

The attributes to use to create the context.

Methods

impl<'a> ContextBuilder<'a>

pub fn new() -> ContextBuilder<'a>

Initializes a new ContextBuilder with default values.

pub fn with_gl(self, request: GlRequest) -> ContextBuilder<'a>

Sets how the backend should choose the OpenGL API and version.

pub fn with_gl_profile(self, profile: GlProfile) -> ContextBuilder<'a>

Sets the desired OpenGL context profile.

pub fn with_gl_debug_flag(self, flag: bool) -> ContextBuilder<'a>

Sets the debug flag for the OpenGL context.

The default value for this flag is cfg!(debug_assertions), which means that it's enabled when you run cargo build and disabled when you run cargo build --release.

pub fn with_gl_robustness(self, robustness: Robustness) -> ContextBuilder<'a>

Sets the robustness of the OpenGL context. See the docs of Robustness.

pub fn with_vsync(self, vsync: bool) -> ContextBuilder<'a>

Requests that the window has vsync enabled.

By default, vsync is not enabled.

pub fn with_shared_lists(self, other: &'a Context) -> ContextBuilder<'a>

Share the display lists with the given Context.

pub fn with_multisampling(self, samples: u16) -> ContextBuilder<'a>

Sets the multisampling level to request. A value of 0 indicates that multisampling must not be enabled.

Panic

Will panic if samples is not a power of two.

pub fn with_depth_buffer(self, bits: u8) -> ContextBuilder<'a>

Sets the number of bits in the depth buffer.

pub fn with_stencil_buffer(self, bits: u8) -> ContextBuilder<'a>

Sets the number of bits in the stencil buffer.

pub fn with_pixel_format(
self,
color_bits: u8,
alpha_bits: u8
) -> ContextBuilder<'a>

Sets the number of bits in the color buffer.

pub fn with_stereoscopy(self) -> ContextBuilder<'a>

Request the backend to be stereoscopic.

pub fn with_srgb(self, srgb_enabled: bool) -> ContextBuilder<'a>

Sets whether sRGB should be enabled on the window.

The default value is false.