Trait conrod::backend::glium::glium::backend::Backend

pub unsafe trait Backend {
    fn swap_buffers(&self) -> Result<(), SwapBuffersError>;

    unsafe fn get_proc_address(&self, symbol: &str) -> *const c_void;

    fn get_framebuffer_dimensions(&self) -> (u32, u32);

    fn is_current(&self) -> bool;

    unsafe fn make_current(&self);
}

Trait for types that can be used as a backend for a glium context.

This trait is unsafe, as you can get undefined behaviors or crashes if you don't implement the methods correctly.

Required Methods

fn swap_buffers(&self) -> Result<(), SwapBuffersError>

Swaps buffers at the end of a frame.

unsafe fn get_proc_address(&self, symbol: &str) -> *const c_void

Returns the address of an OpenGL function.

Supposes that the context has been made current before this function is called.

fn get_framebuffer_dimensions(&self) -> (u32, u32)

Returns the dimensions of the window, or screen, etc.

fn is_current(&self) -> bool

Returns true if the OpenGL context is the current one in the thread.

unsafe fn make_current(&self)

Makes the OpenGL context the current context in the current thread.