Enum conrod::backend::glium::glium::draw_parameters::StencilTest

pub enum StencilTest {
    AlwaysPass,
    AlwaysFail,
    IfLess {
        mask: u32,
    },
    IfLessOrEqual {
        mask: u32,
    },
    IfMore {
        mask: u32,
    },
    IfMoreOrEqual {
        mask: u32,
    },
    IfEqual {
        mask: u32,
    },
    IfNotEqual {
        mask: u32,
    },
}

Specifies which comparison the GPU will do to determine whether a sample passes the stencil test. The general equation is (ref & mask) CMP (stencil & mask), where ref is the reference value (stencil_reference_value_clockwise or stencil_reference_value_counter_clockwise), CMP is the comparison chosen, and stencil is the current value in the stencil buffer.

Variants
AlwaysPass

The stencil test always passes.

AlwaysFail

The stencil test always fails.

IfLess

(ref & mask) < (stencil & mask)

Fields of IfLess

mask: u32

The mask that is and'ed with the reference value and stencil buffer.

IfLessOrEqual

(ref & mask) <= (stencil & mask)

Fields of IfLessOrEqual

mask: u32

The mask that is and'ed with the reference value and stencil buffer.

IfMore

(ref & mask) > (stencil & mask)

Fields of IfMore

mask: u32

The mask that is and'ed with the reference value and stencil buffer.

IfMoreOrEqual

(ref & mask) >= (stencil & mask)

Fields of IfMoreOrEqual

mask: u32

The mask that is and'ed with the reference value and stencil buffer.

IfEqual

(ref & mask) == (stencil & mask)

Fields of IfEqual

mask: u32

The mask that is and'ed with the reference value and stencil buffer.

IfNotEqual

(ref & mask) != (stencil & mask)

Fields of IfNotEqual

mask: u32

The mask that is and'ed with the reference value and stencil buffer.