Module conrod::backend::glium::glium::texture

A texture is an image loaded in video memory, which can be sampled in your shaders.

Texture kinds

One thing that is important to understand when it comes to textures is that the way a texture is accessed (in other words, its "public API") is disconnected from the internal representation of the data.

When it comes to accessing a texture, there are six kinds of textures:

Floating-point textures.

Integral textures (that contain signed integers).

Unsigned textures (that contain unsigned integers).

Depth textures (that contain depth information).

Stencil textures (that contain stencil information).

Depth-stencil textures (that contain at the same time depth and stencil information).

Textures have a different API depending on their kind. For example a integral texture can only be sampled in GLSL through a sampler type which is prefixed with i.

The internal format can only be chosen when the texture is created, and then can never be touched again. Integral and unsigned textures can only contain signed integers and unsigned integers.

Floating-point textures can contain either floating-points or integers. If integers are used, then the maximum value corresponds to the floating-point value 1.0 and the minimal value to 0.0. For example if a texture contains u8s and internally contains the value 128, reading from the texture will yield the value 0.5.

Dimensions

Textures come in nine different dimensions:

Textures with one dimension.

Textures with two dimensions.

Textures with two dimensions and multisampling enabled.

Textures with three dimensions.

Cube textures, which are arrays of six two-dimensional textures corresponding to the six faces of a cube.

Arrays of one-dimensional textures.

Arrays of two-dimensional textures.

Arrays of two-dimensional textures with multisampling enabled.

Arrays of cube textures.

The difference between a 3D texture and a 2D textures array (and between a 2D texture and a 1D textures array) is that texture arrays can only be accessed by individual layers. That is, you can only access layer 0, or layer 1, or layer 2, and so on. Whereas if you use 3D textures you can access layer 0.5 for example.

All textures except depth, stencil and depth-stencil textures have mipmaps. A mipmap is a smaller version of the texture whose purpose is to be used during rendering when the texture will be small on the screen.

Texture types in glium

In addition to the nine different dimensions types, there are nine kinds of texture formats:

The texture contains floating-point data, with either the Compressed prefix or no prefix at all.

The texture contains floating-point data in the sRGB color space, with either the Compressed prefix or not.

The texture contains signed integers, with the Integral prefix.

The texture contains unsigned integers, with the Unsigned prefix.

The texture contains depth information, with the Depth prefix.

The texture contains stencil information, with the Stencil prefix.

The texture contains depth and stencil information, with the DepthStencil prefix.

Each combination of dimensions and format corresponds to a sampler type in GLSL and in glium. For example, an IntegralTexture3d can only be bound to an isampler3D uniform in GLSL.

The difference between compressed textures and uncompressed textures is that you can't do render-to-texture on the former.

The most common types of textures are CompressedSrgbTexture2d, SrgbTexture2d and Texture2d (the two dimensions being the width and height). These are what you will use most of the time.

Buffer textures

A BufferTexture is a special kind of one-dimensional texture that gets its data from a buffer. Buffer textures have very limited capabilities (you can't draw to them for example). They are an alternative to uniform buffers and SSBOs.

See the buffer_textures module for more infos.

About sRGB

For historical reasons, the color data contained in almost all image files are not in RGB but in sRGB. sRGB colors are slightly brighter than linear RGB in order to compensate for the fact that screens darken some values that they receive.

When you load image files, you are encouraged to create sRGB textures (with SrgbTexture2d instead of Texture2d for example).

By default, glium enables the GL_FRAMEBUFFER_SRGB trigger, which expects the output of your fragment shader to be in linear RGB and then turns it into sRGB before writing in the framebuffer. Sampling from an sRGB texture will convert the texture colors from sRGB to RGB. If you create a regular RGB texture and put sRGB data in it, then the result will be too bright.

Bindless textures

Bindless textures are a very recent feature that is supported only by recent hardware and drivers.

Without bindless textures, using a texture in a shader requires binding the texture to a specific bind point before drawing. This not only slows down rendering, but may also prevent you from grouping multiple draw calls into one because of the limitation to the number of available texture units.

Instead, bindless textures allow you to manually manipulate pointers to textures in video memory. You can use thousands of textures if you want.

Modules

bindless
Without bindless textures, using a texture in a shader requires binding the texture to a specific bind point before drawing. This not only slows down rendering, but may also prevent you from grouping multiple draw calls into one because of the limitation to the number of available texture units.

buffer_texture
A BufferTexture is a special kind of one-dimensional texture that gets its data from a buffer. Buffer textures have very limited capabilities compared to other texture types.

compressed_cubemap
Contains the implementation of CompressedCubemap.

compressed_cubemap_array
Contains the implementation of CompressedCubemapArray.

compressed_srgb_cubemap
Contains the implementation of CompressedSrgbCubemap.

compressed_srgb_cubemap_array
Contains the implementation of CompressedSrgbCubemapArray.

compressed_srgb_texture1d
Contains the implementation of CompressedSrgbTexture1d.

compressed_srgb_texture1d_array
Contains the implementation of CompressedSrgbTexture1dArray.

compressed_srgb_texture2d
Contains the implementation of CompressedSrgbTexture2d.

compressed_srgb_texture2d_array
Contains the implementation of CompressedSrgbTexture2dArray.

compressed_srgb_texture3d
Contains the implementation of CompressedSrgbTexture3d.

compressed_texture1d
Contains the implementation of CompressedTexture1d.

compressed_texture1d_array
Contains the implementation of CompressedTexture1dArray.

compressed_texture2d
Contains the implementation of CompressedTexture2d.

compressed_texture2d_array
Contains the implementation of CompressedTexture2dArray.

compressed_texture3d
Contains the implementation of CompressedTexture3d.

cubemap
Contains the implementation of Cubemap.

cubemap_array
Contains the implementation of CubemapArray.

depth_cubemap
Contains the implementation of DepthCubemap.

depth_cubemap_array
Contains the implementation of DepthCubemapArray.

depth_stencil_cubemap
Contains the implementation of DepthStencilCubemap.

depth_stencil_cubemap_array
Contains the implementation of DepthStencilCubemapArray.

depth_stencil_texture1d
Contains the implementation of DepthStencilTexture1d.

depth_stencil_texture1d_array
Contains the implementation of DepthStencilTexture1dArray.

depth_stencil_texture2d
Contains the implementation of DepthStencilTexture2d.

depth_stencil_texture2d_array
Contains the implementation of DepthStencilTexture2dArray.

depth_stencil_texture2d_multisample
Contains the implementation of DepthStencilTexture2dMultisample.

depth_stencil_texture2d_multisample_array
Contains the implementation of DepthStencilTexture2dMultisampleArray.

depth_stencil_texture3d
Contains the implementation of DepthStencilTexture3d.

depth_texture1d
Contains the implementation of DepthTexture1d.

depth_texture1d_array
Contains the implementation of DepthTexture1dArray.

depth_texture2d
Contains the implementation of DepthTexture2d.

depth_texture2d_array
Contains the implementation of DepthTexture2dArray.

depth_texture2d_multisample
Contains the implementation of DepthTexture2dMultisample.

depth_texture2d_multisample_array
Contains the implementation of DepthTexture2dMultisampleArray.

depth_texture3d
Contains the implementation of DepthTexture3d.

integral_cubemap
Contains the implementation of IntegralCubemap.

integral_cubemap_array
Contains the implementation of IntegralCubemapArray.

integral_texture1d
Contains the implementation of IntegralTexture1d.

integral_texture1d_array
Contains the implementation of IntegralTexture1dArray.

integral_texture2d
Contains the implementation of IntegralTexture2d.

integral_texture2d_array
Contains the implementation of IntegralTexture2dArray.

integral_texture2d_multisample
Contains the implementation of IntegralTexture2dMultisample.

integral_texture2d_multisample_array
Contains the implementation of IntegralTexture2dMultisampleArray.

integral_texture3d
Contains the implementation of IntegralTexture3d.

pixel_buffer
Pixel buffers are buffers that contain two-dimensional texture data.

srgb_cubemap
Contains the implementation of SrgbCubemap.

srgb_cubemap_array
Contains the implementation of SrgbCubemapArray.

srgb_texture1d
Contains the implementation of SrgbTexture1d.

srgb_texture1d_array
Contains the implementation of SrgbTexture1dArray.

srgb_texture2d
Contains the implementation of SrgbTexture2d.

srgb_texture2d_array
Contains the implementation of SrgbTexture2dArray.

srgb_texture2d_multisample
Contains the implementation of SrgbTexture2dMultisample.

srgb_texture2d_multisample_array
Contains the implementation of SrgbTexture2dMultisampleArray.

srgb_texture3d
Contains the implementation of SrgbTexture3d.

stencil_cubemap
Contains the implementation of StencilCubemap.

stencil_cubemap_array
Contains the implementation of StencilCubemapArray.

stencil_texture1d
Contains the implementation of StencilTexture1d.

stencil_texture1d_array
Contains the implementation of StencilTexture1dArray.

stencil_texture2d
Contains the implementation of StencilTexture2d.

stencil_texture2d_array
Contains the implementation of StencilTexture2dArray.

stencil_texture2d_multisample
Contains the implementation of StencilTexture2dMultisample.

stencil_texture2d_multisample_array
Contains the implementation of StencilTexture2dMultisampleArray.

texture1d
Contains the implementation of Texture1d.

texture1d_array
Contains the implementation of Texture1dArray.

texture2d
Contains the implementation of Texture2d.

texture2d_array
Contains the implementation of Texture2dArray.

texture2d_multisample
Contains the implementation of Texture2dMultisample.

texture2d_multisample_array
Contains the implementation of Texture2dMultisampleArray.

texture3d
Contains the implementation of Texture3d.

unsigned_cubemap
Contains the implementation of UnsignedCubemap.

unsigned_cubemap_array
Contains the implementation of UnsignedCubemapArray.

unsigned_texture1d
Contains the implementation of UnsignedTexture1d.

unsigned_texture1d_array
Contains the implementation of UnsignedTexture1dArray.

unsigned_texture2d
Contains the implementation of UnsignedTexture2d.

unsigned_texture2d_array
Contains the implementation of UnsignedTexture2dArray.

unsigned_texture2d_multisample
Contains the implementation of UnsignedTexture2dMultisample.

unsigned_texture2d_multisample_array
Contains the implementation of UnsignedTexture2dMultisampleArray.

unsigned_texture3d
Contains the implementation of UnsignedTexture3d.

Structs

BindlessTexturesNotSupportedError
Bindless textures are not supported.

CompressedCubemap
A compressed cube texture containing floating-point data.

CompressedCubemapArray
An array of compressed cube textures containing floating-point data.

CompressedSrgbCubemap
A compressed cube texture containing sRGB floating-point data.

CompressedSrgbCubemapArray
An array of compressed cube textures containing sRGB floating-point data.

CompressedSrgbTexture1d
A compressed one-dimensional texture containing sRGB floating-point data.

CompressedSrgbTexture1dArray
An array of compressed one-dimensional textures containing sRGB floating-point data.

CompressedSrgbTexture2d
A compressed two-dimensional texture containing sRGB floating-point data.

CompressedSrgbTexture2dArray
An array of compressed two-dimensional textures containing sRGB floating-point data.

CompressedSrgbTexture3d
A compressed three-dimensional texture containing sRGB floating-point data.

CompressedTexture1d
A compressed one-dimensional texture containing floating-point data.

CompressedTexture1dArray
An array of compressed one-dimensional textures containing floating-point data.

CompressedTexture2d
A compressed two-dimensional texture containing floating-point data.

CompressedTexture2dArray
An array of compressed two-dimensional textures containing floating-point data.

CompressedTexture3d
A compressed three-dimensional texture containing floating-point data.

Cubemap
A cube texture containing floating-point data.

CubemapArray
An array of cube textures containing floating-point data.

DepthCubemap
A cube texture containing depth data.

DepthCubemapArray
An array of cube textures containing depth data.

DepthStencilCubemap
A cube texture containing both depth and stencil data.

DepthStencilCubemapArray
An array of cube textures containing both depth and stencil data.

DepthStencilTexture1d
A one-dimensional texture containing both depth and stencil data.

DepthStencilTexture1dArray
An array of one-dimensional textures containing both depth and stencil data.

DepthStencilTexture2d
A two-dimensional texture containing both depth and stencil data.

DepthStencilTexture2dArray
An array of two-dimensional textures containing both depth and stencil data.

DepthStencilTexture2dMultisample
A two-dimensional texture containing both depth and stencil data.

DepthStencilTexture2dMultisampleArray
An array of two-dimensional textures containing both depth and stencil data.

DepthStencilTexture3d
A three-dimensional texture containing both depth and stencil data.

DepthTexture1d
A one-dimensional texture containing depth data.

DepthTexture1dArray
An array of one-dimensional textures containing depth data.

DepthTexture2d
A two-dimensional texture containing depth data.

DepthTexture2dArray
An array of two-dimensional textures containing depth data.

DepthTexture2dMultisample
A two-dimensional texture containing depth data.

DepthTexture2dMultisampleArray
An array of two-dimensional textures containing depth data.

DepthTexture3d
A three-dimensional texture containing depth data.

IntegralCubemap
A cube texture containing signed integral data.

IntegralCubemapArray
An array of cube textures containing signed integral data.

IntegralTexture1d
A one-dimensional texture containing signed integral data.

IntegralTexture1dArray
An array of one-dimensional textures containing signed integral data.

IntegralTexture2d
A two-dimensional texture containing signed integral data.

IntegralTexture2dArray
An array of two-dimensional textures containing signed integral data.

IntegralTexture2dMultisample
A two-dimensional texture containing signed integral data.

IntegralTexture2dMultisampleArray
An array of two-dimensional textures containing signed integral data.

IntegralTexture3d
A three-dimensional texture containing signed integral data.

RawImage1d
Represents raw data for a two-dimensional image.

RawImage2d
Represents raw data for a two-dimensional image.

RawImage3d
Represents raw data for a two-dimensional image.

ResidentTexture
A texture that is resident in video memory. This allows you to use bindless textures in your shaders.

SrgbCubemap
A cube texture containing sRGB floating-point data.

SrgbCubemapArray
An array of cube textures containing sRGB floating-point data.

SrgbTexture1d
A one-dimensional texture containing sRGB floating-point data.

SrgbTexture1dArray
An array of one-dimensional textures containing sRGB floating-point data.

SrgbTexture2d
A two-dimensional texture containing sRGB floating-point data.

SrgbTexture2dArray
An array of two-dimensional textures containing sRGB floating-point data.

SrgbTexture2dMultisample
A two-dimensional texture containing sRGB floating-point data.

SrgbTexture2dMultisampleArray
An array of two-dimensional textures containing sRGB floating-point data.

SrgbTexture3d
A three-dimensional texture containing sRGB floating-point data.

StencilCubemap
A cube texture containing stencil data.

StencilCubemapArray
An array of cube textures containing stencil data.

StencilTexture1d
A one-dimensional texture containing stencil data.

StencilTexture1dArray
An array of one-dimensional textures containing stencil data.

StencilTexture2d
A two-dimensional texture containing stencil data.

StencilTexture2dArray
An array of two-dimensional textures containing stencil data.

StencilTexture2dMultisample
A two-dimensional texture containing stencil data.

StencilTexture2dMultisampleArray
An array of two-dimensional textures containing stencil data.

Texture1d
A one-dimensional texture containing floating-point data.

Texture1dArray
An array of one-dimensional textures containing floating-point data.

Texture2d
A two-dimensional texture containing floating-point data.

Texture2dArray
An array of two-dimensional textures containing floating-point data.

Texture2dMultisample
A two-dimensional texture containing floating-point data.

Texture2dMultisampleArray
An array of two-dimensional textures containing floating-point data.

Texture3d
A three-dimensional texture containing floating-point data.

TextureAny
A texture whose type isn't fixed at compile-time.

TextureAnyImage
Represents a specific 2D image of a texture. 1D textures are considered as having a height of 1.

TextureAnyLayer
Represents a specific layer of an array texture and 3D textures.

TextureAnyLayerMipmap
Represents a specific layer of a specific mipmap. This is the same as TextureAnyImage, except for 3D textures, cubemaps and cubemap arrays.

TextureAnyMipmap
Represents a specific mipmap of a texture.

TextureHandle
Represents a handle to a texture. Contains a raw pointer to a texture that is hidden from you.

UnsignedCubemap
A cube texture containing unsigned integral data.

UnsignedCubemapArray
An array of cube textures containing unsigned integral data.

UnsignedTexture1d
A one-dimensional texture containing unsigned integral data.

UnsignedTexture1dArray
An array of one-dimensional textures containing unsigned integral data.

UnsignedTexture2d
A two-dimensional texture containing unsigned integral data.

UnsignedTexture2dArray
An array of two-dimensional textures containing unsigned integral data.

UnsignedTexture2dMultisample
A two-dimensional texture containing unsigned integral data.

UnsignedTexture2dMultisampleArray
An array of two-dimensional textures containing unsigned integral data.

UnsignedTexture3d
A three-dimensional texture containing unsigned integral data.

Enums

ClientFormat
List of client-side pixel formats.

CompressedFormat
List of compressed texture formats.

CompressedMipmapsOption
Describes what to do about mipmaps during compressed texture creation.

CompressedSrgbFormat
List of compressed pixel formats in the sRGB color space.

CubeLayer
Represents a layer of a cubemap.

DepthFormat
List of formats available for depth textures.

DepthStencilFormat
List of formats available for depth-stencil textures.

Dimensions
Type of a texture.

GetFormatError
Error that can happen when retrieving the internal format of a texture.

InternalFormat
Internal format of a texture.

InternalFormatType
Format of a component of an internal format.

MipmapsOption
Describes what to do about mipmaps during texture creation.

SrgbFormat
List of uncompressed pixel formats that contain floating-point data in the sRGB color space.

StencilFormat
List of formats available for stencil textures.

TextureCreationError
Error that can happen when creating a texture.

TextureFormat
Format of the internal representation of a texture.

TextureKind
Represents a kind of texture.

UncompressedFloatFormat
List of uncompressed pixel formats that contain floating-point-like data.

UncompressedIntFormat
List of uncompressed pixel formats that contain signed integral data.

UncompressedUintFormat
List of uncompressed pixel formats that contain unsigned integral data.

Traits

PixelValue
A trait that must be implemented for any type that can represent the value of a pixel.

Texture1dDataSink
Trait that describes types that can be built from one-dimensional texture data.

Texture1dDataSource
Trait that describes data for a one-dimensional texture.

Texture2dDataSink
Trait that describes types that can be built from two-dimensional texture data.

Texture2dDataSource
Trait that describes data for a two-dimensional texture.

Texture3dDataSink
Trait that describes types that can be built from one-dimensional texture data.

Texture3dDataSource
Trait that describes data for a two-dimensional texture.

ToClientFormat

Functions

is_cubemap_arrays_supported
Returns true is cubemap arrays are supported.

is_cubemaps_supported
Returns true is cubemaps are supported.

is_texture_1d_array_supported
Returns true is one-dimensional texture arrays are supported.

is_texture_1d_supported
Returns true is one-dimensional textures are supported.

is_texture_2d_array_supported
Returns true is two-dimensional texture arrays are supported.

is_texture_2d_multisample_array_supported
Returns true is two-dimensional multisample texture arrays are supported.

is_texture_2d_multisample_supported
Returns true is two-dimensional multisample textures are supported.

is_texture_2d_supported
Returns true is two-dimensional textures are supported.

is_texture_3d_supported
Returns true is three-dimensional textures are supported.