Enum conrod::backend::glium::glium::texture::TextureCreationError

pub enum TextureCreationError {
    FormatNotSupported,
    DimensionsNotSupported,
    TypeNotSupported,
}

Error that can happen when creating a texture.

Variants
FormatNotSupported

The requested format is not supported by the backend.

DimensionsNotSupported

The requested texture dimensions are not supported.

TypeNotSupported

The texture format is not supported by the backend.