Struct conrod::backend::glium::Renderer

pub struct Renderer { /* fields omitted */ }

A type used for translating render::Primitives into Commands that indicate how to draw the conrod GUI using glium.

Methods

impl Renderer

pub fn new<F>(facade: &F) -> Result<Self, RendererCreationError> where
F: Facade,

Construct a new empty Renderer.

Important traits for Commands<'a>
impl<'a> Iterator for Commands<'a> type Item = Command<'a>;

pub fn commands(&self) -> Commands

Produce an Iterator yielding Commands.

pub fn fill<P, T>(
&mut self,
display: &Display,
primitives: P,
image_map: &Map<T>
) where
P: PrimitiveWalker,
T: TextureDimensions,

Fill the inner vertex and command buffers by translating the given primitives.

pub fn draw<F, S, T>(
&self,
facade: &F,
surface: &mut S,
image_map: &Map<T>
) -> Result<(), DrawError> where
F: Facade,
S: Surface,
Sampler<'a, T>: AsUniformValue,

Draws using the inner list of Commands to the given display.

Note: If you require more granular control over rendering, you may want to use the fill and commands methods separately. This method is simply a convenience wrapper around those methods for the case that the user does not require accessing or modifying conrod's draw parameters, uniforms or generated draw commands.