Module conrod::backend::graphics

Conrod's generic graphics backend.

Note: Conrod currently uses Piston's generic graphics crate (and specifically the Graphics) and CharacterCache traits to enable genericity over custom user backends. This dependency may change in the near future in favour of a simplified conrod-specific graphics and character caching backend trait.

Structs

Character
Holds rendered character data.

Context
Drawing 2d context.

DrawState
An assembly of states that affect regular draw calls

Traits

CharacterCache
Stores characters in a buffer and loads them by demand.

Graphics
Implemented by all graphics back-ends.

ImageSize
Implemented by all images to be used with generic algorithms.

Transformed
Implemented by contexts that can transform.

Functions

draw_from_container
Use the given CharacterCache and Graphics backends to draw the given widget.

draw_from_graph
Draw the given Graph using the given CharacterCache and Graphics backends.

draw_lines
Draw a series of lines between the given Points using the given style.

draw_rectangle
Draw a rectangle at the given Rect.

draw_scrolling
Draw the scroll bars (if necessary) for the given widget's scroll state.