Module conrod::backend::piston::draw

A piston backend for rendering conrod primitives.

Structs

Context
Drawing 2d context.

DrawState
Graphics draw state used for blending, clipping and stencil rendering.

Traits

Graphics
Implemented by all graphics back-ends.

ImageSize
Implemented by all images to be used with generic algorithms.

Transformed
Implemented by contexts that can transform.

Functions

primitive
Render a single Primitive.

primitives
Render the given sequence of conrod primitive widgets.