Module conrod::backend::piston::gfx

Contains the GfxContext wrapper for convenient integration with backend::piston::Window

Structs

Context
Drawing 2d context.

DrawState
Graphics draw state used for blending, clipping and stencil rendering.

GfxContext
Contains state used by Gfx to draw. Can be stored within a window.

GlyphCache
A wrapper around a G2dTexture and a rusttype GPU Cache

Texture
Represents a texture.

TextureSettings
Texture creation parameters.

Enums

Flip
Flip settings.

GlyphError
An enum to represent various possible run-time errors that may occur.

Traits

Graphics
Implemented by all graphics back-ends.

ImageSize
Implemented by all images to be used with generic algorithms.

Transformed
Implemented by contexts that can transform.

Functions

draw
Renders the given sequence of conrod primitives.

draw_primitive
Draw a single Primitive to the screen.

Type Definitions

G2d
2D graphics.

G2dTexture
Texture type compatible with G2d.

GfxEncoder
Actual gfx::Stream implementation carried by the window.

Glyphs
Glyph cache.