Struct conrod::backend::piston::gfx::GlyphCache

pub struct GlyphCache { /* fields omitted */ }

A wrapper around a G2dTexture and a rusttype GPU Cache

Using a wrapper simplifies the construction of both caches and ensures that they maintain identical dimensions.

Methods

impl GlyphCache

pub fn new(window: &mut Window, width: u32, height: u32) -> GlyphCache

Constructor for a new GlyphCache, using the factory associated with a Window

impl GlyphCache

pub fn new_from_factory(
factory: &mut Factory,
width: u32,
height: u32
) -> GlyphCache

Constructor for a new GlyphCache.

The width and height arguments are in pixel values.

If you need to resize the GlyphCache, construct a new one and discard the old one.