Struct conrod::backend::piston::window::Window

pub struct Window<W: BasicWindow = GlutinWindow> {
    pub window: W,
    pub context: GfxContext,
}

Contains everything required for controlling window, graphics, event loop.

Fields
window: W

The window.

context: GfxContext

Stores state associated with Gfx.

Methods

impl<W> Window<W> where W: BasicWindow, W::Event: GenericEvent

fn new(window: W, context: GfxContext) -> Self where W: OpenGLWindow

Creates a new piston window.

fn draw_2d<E, F, U>(&mut self, e: &E, f: F) -> Option<U> where W: OpenGLWindow, E: GenericEvent, F: FnOnce(Context, &mut G2d) -> U

Renders 2D graphics.

fn next_event<E>(&mut self, events: &mut E) -> Option<Event> where Self: EventWindow<E>

A wrapper around the EventWindow::next trait method. This avoids the need for users to import the EventWindow trait in most cases.

fn handle_event(&mut self, event: &Event<W::Event>)

Let window handle new event.