Struct conrod::backend::piston_window::GlyphCache

pub struct GlyphCache { /* fields omitted */ }

A wrapper around a G2dTexture and a rusttype GPU Cache

Using a wrapper simplifies the construction of both caches and ensures that they maintain identical dimensions.

Methods

impl GlyphCache

fn new<B>(window: &mut PistonWindow<B>, width: u32, height: u32) -> GlyphCache where B: Window

Constructor for a new GlyphCache.

The PistonWindow provides the Factory argument to the G2dTexture constructor.

The width and height arguments are in pixel values.

If you need to resize the GlyphCache, construct a new one and discard the old one.