Module conrod::cursor

Contains an extendable enum of supported mouse cursor types.

Use this module to map from the conrod's mouse cursor types to the types known to the window backend you are using. A lot of these are already implemented in conrod::backend. Unless you are using custom mouse cursor types not provided here, then using one of the implementations in conrod::backend should be sufficient.

Enums

MouseCursor
This enum specifies cursor types used by internal widgets. For custom widgets using custom cursor types, you can still use this enum by specifying a numbered custom variant.