Enum conrod::Position

pub enum Position {
    Absolute(Scalar),
    Relative(Scalar, Option<Index>),
    Align(Align, Option<Index>),
    Direction(Direction, Scalar, Option<Index>),
    Place(Place, Option<Index>),
}

Some Position of some Widget along a single axis.

Positions for both the x and y axes are stored internally within the widget::CommonBuilder type, allowing all widgets to be positioned in a variety of different ways.

See the Positionable trait for methods that allow for setting the Positions in various ways.

Note that Positionable is implemented for all types that implement Widget.

Variants

Absolute

A specific position.

Relative

A position relative to some other Widget.

Align

A position aligned with some other Widget.

Direction

A direction relative to some other Widget.

Place

A position at a place on some other Widget.