Enum conrod::event::Button

pub enum Button {
    Keyboard(Key),
    Mouse(MouseButton, Point),
    Controller(ControllerButton),
}

The different kinds of Buttons that may be Pressed or Released.

Variants
Keyboard(Key)

A keyboard button.

Mouse(MouseButton, Point)

A mouse button along with the location at which it was Pressed/Released.

Controller(ControllerButton)

A controller button.

Methods

impl Button

fn relative_to(&self, xy: Point) -> Button

Returns a copy of the Button relative to the given xy