Enum conrod::event::Input

pub enum Input {
    Press(Button),
    Release(Button),
    Resize(u32, u32),
    Move(Motion),
    Text(String),
    Focus(bool),
}

The event type that is used by conrod to track inputs from the world. Events yielded by polling window backends should be converted to this type. This can be thought of as the event type which is supplied by the window backend to drive the state of the Ui forward.

This type is solely used within the Ui::handle_event method. The Input events are interpreted to create higher level Events (such as DoubleClick, WidgetCapturesKeyboard, etc) which are stored for later processing by Widgets, which will occur during the call to Ui::set_widgets.

Note: Input events that contain co-ordinates must be oriented with (0, 0) at the middle of the window with the y axis pointing upwards (Cartesian co-ordinates). All co-ordinates and dimensions must be given as Scalar (DPI agnostic) values. Many windows provide coordinates with the origin in the top left with y pointing down, so you might need to translate these co-ordinates when converting to this event. Also be sure to invert the y axis of MouseScroll events.

Variants
Press(Button)

A button on some input device was pressed.

Release(Button)

A button on some input device was released.

Resize(u32, u32)

The window was received to the given dimensions.

Move(Motion)

Some motion input was received (e.g. moving mouse, joystick or touch pad).

Text(String)

Text input was received, usually via the keyboard.

Focus(bool)

The window was focused or lost focus.