Enum conrod::event::Ui

pub enum Ui {
    Text(Option<Id>, Text),
    Press(Option<Id>, Press),
    Release(Option<Id>, Release),
    Move(Option<Id>, Move),
    WindowResized(Dimensions),
    Click(Option<Id>, Click),
    DoubleClick(Option<Id>, DoubleClick),
    Drag(Option<Id>, Drag),
    Scroll(Option<Id>, Scroll),
    WidgetCapturesMouse(Id),
    WidgetUncapturesMouse(Id),
    WidgetCapturesKeyboard(Id),
    WidgetUncapturesKeyboard(Id),
}

Represents all events interpreted by the Ui.

Variants
Text(Option<Id>, Text)

Entered text, along with the widget that was capturing the keyboard at the time.

Press(Option<Id>, Press)

Some button was pressed, along with the widget that was capturing the device whose button which was pressed.

Release(Option<Id>, Release)

Some button was released, along with the widget that was capturing the device whose button which was released.

Move(Option<Id>, Move)

Represents all forms of motion input, alongside with the widget that was capturing the mouse at the time.

WindowResized(Dimensions)

The window's dimensions were resized.

Click(Option<Id>, Click)

Represents a pointing device being pressed and subsequently released while over the same location.

DoubleClick(Option<Id>, DoubleClick)

Two Click events with the same button and xy occurring within a duration that is less that the theme.double_click_threshold.

Drag(Option<Id>, Drag)

Represents a pointing device button being pressed and a subsequent movement of the mouse.

Scroll(Option<Id>, Scroll)

A generic scroll event.

Scroll does not necessarily have to get created by a mouse wheel, it could be generated from a keypress, or as a response to handling some other event.

Received Scroll events are first applied to all scrollable widgets under the mouse from top to bottom. The remainder will then be applied to either 1. whatever widget captures the device from which the scroll was emitted or 2. whatever widget was specified.

WidgetCapturesMouse(Id)

Indicates that the given widget has captured the mouse.

WidgetUncapturesMouse(Id)

Indicates that the given widget has released the mouse from capturing.

WidgetCapturesKeyboard(Id)

Indicates that the given widget has captured the keyboard.

WidgetUncapturesKeyboard(Id)

Indicates that the given widget has released the keyboard from capturing.