conrod::event::Raw

type Raw = Event<Input>;

The event type that is used by conrod to track inputs from the world. Events yielded by polling window backends should be converted to this type. This can be thought of as the event type which is supplied by the window backend to drive the state of the Ui forward.

This type is solely used within the Ui::handle_event method which immediately converts given user events to RawEvents via the ToRawEvent trait bound. The RawEvents are interpreted to create higher level Events (such as DoubleClick, WidgetCapturesKeyboard, etc) which are stored for later processing by Widgets, which will occur during the call to Ui::set_widgets.

Note: Raw events that contain co-ordinates must be oriented with (0, 0) at the middle of the window with the y axis pointing upwards (Cartesian co-ordinates). Many windows provide coordinates with the origin in the top left with y pointing down, so you might need to translate these co-ordinates when implementing this method. Also be sure to invert the y axis of MouseScroll events.