Module conrod::events::global_input

Handles all of the global input events and state. The core of this module is the GlobalInput struct. It is responsible for aggregating and interpreting raw input events into high-level semantic events.

Structs

GlobalInput
Global input event handler that also implements InputProvider. The Ui passes all events to it's GlobalInput instance, which aggregates and interprets the events to provide so-called 'high-level' events to widgets. This input gets reset after every update by the Ui.

Type Definitions

GlobalInputEventIterator
Iterator over global UiEvents. Unlike the WidgetInputEventIterator, this will never filter out any events, and all coordinates will be reative to the (0,0) origin of the window.