Struct conrod::events::global_input::GlobalInput

pub struct GlobalInput {
    pub start_state: InputState,
    pub current_state: InputState,
    // some fields omitted
}

Global input event handler that also implements InputProvider. The Ui passes all events to it's GlobalInput instance, which aggregates and interprets the events to provide so-called 'high-level' events to widgets. This input gets reset after every update by the Ui.

Fields

start_state

The InputState as it was at the end of the last update cycle.

current_state

The most recent InputState, with updates from handling all the events this update cycle

Methods

impl GlobalInput

fn new(drag_threshold: Scalar) -> GlobalInput

Returns a fresh new GlobalInput

fn push_event(&mut self, event: UiEvent)

Adds a new event and updates the internal state.

fn reset(&mut self)

Called at the end of every update cycle in order to prepare the GlobalInput to handle events for the next one.

fn mouse_position(&self) -> Point

Returns the most up to date position of the mouse

fn starting_state(&self) -> &InputState

Returns the input state as it was after the last update

fn currently_capturing_mouse(&self) -> Option<Index>

Returns the most up to date info on which widget is capturing the mouse

fn currently_capturing_keyboard(&self) -> Option<Index>

Returns the most up to date info on which widget is capturing the keyboard