Trait conrod::events::input_provider::InputProvider

pub trait InputProvider<'a> {
    type Events: Iterator<Item=&'a UiEvent>;
    fn all_events(&'a self) -> Self::Events;
    fn current_state(&'a self) -> &'a InputState;
    fn mouse_button_down(&'a self, button: MouseButton) -> Option<Point>;

    fn text_just_entered(&'a self) -> TextJustEntered<'a, Self::Events> { ... }
    fn keys_just_released(&'a self) -> KeysJustReleased<'a, Self::Events> { ... }
    fn keys_just_pressed(&'a self) -> KeysJustPressed<'a, Self::Events> { ... }
    fn mouse_buttons_just_pressed(&'a self) -> MouseButtonsJustPressed<'a, Self::Events> { ... }
    fn mouse_buttons_just_released(&'a self) -> MouseButtonsJustReleased<'a, Self::Events> { ... }
    fn scroll(&'a self) -> Option<Scroll> { ... }
    fn mouse_left_drag(&'a self) -> Option<MouseDrag> { ... }
    fn mouse_drag(&'a self, button: MouseButton) -> Option<MouseDrag> { ... }
    fn mouse_left_click(&'a self) -> Option<MouseClick> { ... }
    fn mouse_right_click(&'a self) -> Option<MouseClick> { ... }
    fn mouse_click(&'a self, button: MouseButton) -> Option<MouseClick> { ... }
    fn mouse_left_button_down(&'a self) -> Option<Point> { ... }
    fn mouse_right_button_down(&'a self) -> Option<Point> { ... }
    fn mouse_position(&'a self) -> Point { ... }
}

Trait for something that provides events to be consumed by a widget.

Provides a bunch of convenience methods for filtering out specific types of events.

Associated Types

type Events: Iterator<Item=&'a UiEvent>

An iterator yielding references to the InputProvider's UiEvents.

Required Methods

fn all_events(&'a self) -> Self::Events

This is the only method that needs to be implemented. Just provided a reference to a Vec<UiEvent> that contains all the events for this update cycle.

fn current_state(&'a self) -> &'a InputState

Returns the current input state. The returned state is assumed to be up to date with all of the events so far.

fn mouse_button_down(&'a self, button: MouseButton) -> Option<Point>

If the given mouse button is currently pressed, returns the current position of the mouse. Otherwise, returns None

Provided Methods

fn text_just_entered(&'a self) -> TextJustEntered<'a, Self::Events>

Returns a reference to each slice of Text that was entered since the last update.

fn keys_just_released(&'a self) -> KeysJustReleased<'a, Self::Events>

Returns all of the Keys that were released since the last update.

fn keys_just_pressed(&'a self) -> KeysJustPressed<'a, Self::Events>

Returns all of the keyboard Keys that were pressed since the last update.

fn mouse_buttons_just_pressed(&'a self) -> MouseButtonsJustPressed<'a, Self::Events>

Returns all of the MouseButtons that were pressed since the last update.

fn mouse_buttons_just_released(&'a self) -> MouseButtonsJustReleased<'a, Self::Events>

Returns all of the MouseButtons that were released since the last update.

fn scroll(&'a self) -> Option<Scroll>

Returns a Scroll struct if any scrolling was done since the last update. If multiple raw scroll events occured since the last update (which could very well happen if the user is scrolling quickly), then the Scroll returned will represent an aggregate total of all the scrolling.

fn mouse_left_drag(&'a self) -> Option<MouseDrag>

Convenience method to call mouse_drag, passing in MouseButton::Left. Saves widgets from having to use input::mouse::MouseButton if all they care about is the left mouse button.

fn mouse_drag(&'a self, button: MouseButton) -> Option<MouseDrag>

Returns a MouseDrag if one has occured involving the given mouse button. If multiple raw mouse movement events have occured since the last update (which will happen if the user moves the mouse quickly), then the returned MouseDrag will be only the most recent one, which will contain the most recent mouse position.

fn mouse_left_click(&'a self) -> Option<MouseClick>

Convenience method to call mouse_click, passing in passing in MouseButton::Left. Saves widgets from having to use input::mouse::MouseButton if all they care about is the left mouse button.

fn mouse_right_click(&'a self) -> Option<MouseClick>

Convenience method to call mouse_click, passing in passing in MouseButton::Right. Saves widgets from having to use input::mouse::MouseButton if all they care about is the left mouse button.

fn mouse_click(&'a self, button: MouseButton) -> Option<MouseClick>

Returns a MouseClick if one has occured with the given mouse button. A click is determined to have occured if a mouse button was pressed and subsequently released while the mouse was in roughly the same place.

fn mouse_left_button_down(&'a self) -> Option<Point>

Convenience method for checking if the Left mouse button is down. Returns mouse position if the Left mouse button is currently pressed, otherwise None.

fn mouse_right_button_down(&'a self) -> Option<Point>

Convenience method for checking if the Right mouse button is down. Returns mouse position if the Right mouse button is currently pressed, otherwise None.

fn mouse_position(&'a self) -> Point

Convenience method for returning the current mouse position.