Struct conrod::events::input_state::ButtonMap

pub struct ButtonMap {
    // some fields omitted
}

Stores the state of all mouse buttons. If the mouse button is down, it stores the position of the mouse when the button was pressed

Methods

impl ButtonMap

fn new() -> ButtonMap

Returns a new button map with all states set to None

fn set(&mut self, button: MouseButton, point: ButtonDownPosition)

Sets the state of a specific MouseButton

fn get(&self, button: MouseButton) -> ButtonDownPosition

Returns the state of a mouse button

fn take(&mut self, button: MouseButton) -> ButtonDownPosition

Returns the current state of a mouse button, leaving None in its place

fn pressed_button(&self) -> Option<(MouseButton, Point)>

If any mouse buttons are currently pressed, will return a tuple containing both the MouseButton that is pressed and the Point describing the location of the mouse when it was pressed.

fn relative_to(&self, xy: Point) -> ButtonMap

Returns a copy of the ButtonMap relative to the given Point