Struct conrod::events::input_state::InputState

pub struct InputState {
    pub mouse_buttons: ButtonMap,
    pub mouse_position: Point,
    pub widget_capturing_keyboard: Option<Index>,
    pub widget_capturing_mouse: Option<Index>,
    pub modifiers: ModifierKey,
}

Holds the current state of user input. This includes the state of all buttons on the keyboard and mouse, as well as the position of the mouse. It also includes which widgets, if any, are capturing keyboard and mouse input.

Fields

mouse_buttons

A map that stores the up/down state of each button. If the button is down, then it stores the position of the mouse when the button was first pressed.

mouse_position

The current position of the mouse.

widget_capturing_keyboard

Which widget, if any, is currently capturing the keyboard

widget_capturing_mouse

Which widget, if any, is currently capturing the mouse

modifiers

Which modifier keys are being held down.

Methods

impl InputState

fn new() -> InputState

Returns a fresh new input state

fn update(&mut self, event: &UiEvent)

Updates the input state based on an event.

fn relative_to(&self, xy: Point) -> InputState

Returns a copy of the InputState relative to the given position::Point