Enum conrod::events::ui_event::UiEvent

pub enum UiEvent {
    Raw(Input),
    MouseClick(MouseClick),
    MouseDrag(MouseDrag),
    Scroll(Scroll),
    WidgetCapturesMouse(Index),
    WidgetUncapturesMouse(Index),
    WidgetCapturesKeyboard(Index),
    WidgetUncapturesKeyboard(Index),
}

Enum containing all the events that Widgets can listen for.

Variants

Raw

Represents a raw input::Input event

MouseClick

Represents a mouse button being pressed and subsequently released while the mouse stayed in roughly the same place.

MouseDrag

Represents a mouse button being pressed and a subsequent movement of the mouse.

Scroll

This is a generic scroll event. This is different from the input::Movement::MouseScroll event in several aspects. For one, it does not necessarily have to get created by a mouse wheel, it could be generated from a keypress, or as a response to handling some other event. Secondly, it contains a field holding the input::keyboard::ModifierKey that was held while the scroll occured.

WidgetCapturesMouse

Indicates that the given widget is starting to capture the mouse.

WidgetUncapturesMouse

Indicates that the given widget is losing mouse capture.

WidgetCapturesKeyboard

Indicates that the given widget is starting to capture the keyboard.

WidgetUncapturesKeyboard

Indicates that the given widget is losing keyboard capture.

Methods

impl UiEvent

fn relative_to(&self, xy: Point) -> Self

Returns a copy of the UiEvent relative to the given position::Point

fn is_mouse_event(&self) -> bool

Returns true if this event is related to the mouse. Note that just because this method returns true does not mean that the event necessarily came from the mouse. A UiEvent::Scroll is considered to be both a mouse and a keyboard event.

fn is_keyboard_event(&self) -> bool

Returns true if this event is related to the keyboard. Note that just because this method returns true does not mean that the event necessarily came from the keyboard. A UiEvent::Scroll is considered to be both a mouse and a keyboard event.