Struct conrod::events::ui_event::MouseDrag

pub struct MouseDrag {
    pub button: MouseButton,
    pub start: Point,
    pub end: Point,
    pub modifier: ModifierKey,
    pub in_progress: bool,
}

Contains all the relevant information for a mouse drag.

Fields

button

Which mouse button was being held during the drag

start

The origin of the drag. This will always be the position of the mouse whenever the button was first pressed

end

The end position of the mouse. If in_progress is true, then subsequent MouseDrag events may be created with a new end as the mouse continues to move.

modifier

Which modifier keys are being held during the mouse drag.

in_progress

Indicates whether the mouse button is still being held down. If it is, then in_progress will be true and more MouseDrag events can likely be expected.

Methods

impl MouseDrag

fn relative_to(&self, xy: Point) -> MouseDrag

Returns a copy of the MouseDrag relative to the given position::Point