Module conrod::events::widget_input

Contains all the logic for filtering input events and making them relative to widgets. The core of this module is the WidgetInput::for_widget method, which creates an InputProvider that provides input events for a specific widget.

Structs

WidgetInput
Holds any events meant to be given to a Widget. This is what widgets will interface with when handling events in their update method. All events returned from methods on WidgetInput will be relative to the widget's own (0,0) origin. Additionally, WidgetInput will not provide mouse or keyboard events that do not directly pertain to the widget.

WidgetInputEventIterator
Alows iterating over events for a specific widget. All events provided by this Iterator will be filtered, so that input intended for other widgets is excluded. In addition, all mouse events will have their coordinates relative to the widget's own (0,0) origin.