Struct conrod::events::widget_input::WidgetInput

pub struct WidgetInput<'a> {
    // some fields omitted
}

Holds any events meant to be given to a Widget. This is what widgets will interface with when handling events in their update method. All events returned from methods on WidgetInput will be relative to the widget's own (0,0) origin. Additionally, WidgetInput will not provide mouse or keyboard events that do not directly pertain to the widget.

Methods

impl<'a> WidgetInput<'a>

fn for_widget<'g>(widget: Index, widget_area: Rect, global_input: &'g GlobalInput) -> WidgetInput<'g>

Returns a WidgetInput with events specifically for the given widget. Filters out only the events that directly pertain to the widget. All events will also be made relative to the widget's own (0,0) origin.

fn mouse_is_over_widget(&self) -> bool

Returns true if the mouse is currently over the widget, otherwise false

fn maybe_mouse_position(&self) -> Option<Point>

If the mouse is over the widget and no other widget is capturing the mouse, then this will return the position of the mouse relative to the widget. Otherwise, it will return None