Function conrod::mid_left_of

pub fn mid_left_of(target: Dimensions, dim: Dimensions) -> Point

The position of a rect with dim Dimensions at the middle of the inside of the left edge of the target Dimensions.