Module conrod::glyph_cache

Conrod's character caching API.

Provides types and functionality related character texture caching and text dimensioning.

Structs

CharWidths
An iterator yielding the widths of each consecutive character in some sequence.

CharXs
An iterator that converts the given sequnce of chars into their consecutive positions along the x-axis.

GlyphCache
A wrapper over some CharacterCache, exposing it's functionality via a RefCell.

LineBreaksBy
An iterator that yields the indices at which some text should wrap in accordance with the given wrap function.

Lines
A wrapper over an iterator yielding LineBreaks that yields each line divided by the breaks.

Widths
An iterator yielding the widths of each consecutive &str in some sequence.

Enums

LineBreak
The two types of LineBreak indices returned by the WrapIndicesBy iterators.

Type Definitions

HalfW
The half of the width of some character (used within CharXs).

LineBreaksByCharacter
An iterator that yields the indices at which some text should wrap via a character.

LineBreaksByWhitespace
An iterator that yields the indices at which some text should wrap via whitespace.

LinesWrappedBy
An iterator yielding lines for text wrapped with the given function.

LinesWrappedByCharacter
An iterator yielding lines for text wrapped via the first character exceeding a max width.

LinesWrappedByWhitespace
An iterator yielding lines for text wrapped via the first character exceeding a max width.

NextLineBreakFn
A function that returns the first index at which the text should wrap for the given max width.

X
Some position along the X axis (used within CharXs).

Implementations