Struct conrod::glyph_cache::GlyphCache

pub struct GlyphCache<C> {
    // some fields omitted
}

A wrapper over some CharacterCache, exposing it's functionality via a RefCell.

The GlyphCache is frequently needed in order to calculate text dimensions. We wrap the CharacterCache in a RefCell in order to avoid ownership issues that this may cause.

Methods

impl<C> GlyphCache<C>

fn new(cache: C) -> Self

Construct a new GlyphCache.

fn char_width(&self, font_size: FontSize, ch: char) -> Scalar where C: CharacterCache

The width of a single character with the given size.

fn char_widths<I>(&self, font_size: FontSize, chars: I) -> CharWidths<C, I::IntoIter> where I: IntoIterator<Item=char>

Converts the given sequence of chars into their Scalar widths.

fn char_xs<I>(&self, font_size: FontSize, start_x: Scalar, chars: I) -> CharXs<C, I::IntoIter> where C: CharacterCache, I: IntoIterator<Item=char>

Converts the given sequnce of chars into their consecutive positions along the x-axis.

fn width(&self, font_size: FontSize, text: &str) -> Scalar where C: CharacterCache

Return the width of the given text.

fn widths<I>(&self, font_size: FontSize, strs: I) -> Widths<C, I::IntoIter> where I: IntoIterator<Item=&'a str>

Converts the given sequence of &strs into their Scalar widths.

fn line_breaks_by<'a, F>(&'a self, font_size: FontSize, text: &'a str, max_width: Scalar, line_break_fn: F) -> LineBreaksBy<'a, C, F>

An iterator that yields the indices at which some text should wrap in accordance with the given wrap function.

fn line_breaks_by_character<'a>(&'a self, font_size: FontSize, text: &'a str, max_width: Scalar) -> LineBreaksByCharacter<'a, C> where C: CharacterCache

An iterator that yields the indices at which some text should wrap via a character.

fn line_breaks_by_whitespace<'a>(&'a self, font_size: FontSize, text: &'a str, max_width: Scalar) -> LineBreaksByWhitespace<'a, C> where C: CharacterCache

An iterator that yields the indices at which some text should wrap via whitespace.

fn lines_wrapped_by<'a, F>(&'a self, font_size: FontSize, text: &'a str, max_width: Scalar, wrap_fn: F) -> LinesWrappedBy<'a, C, F>

An iterator that behaves the same as text.lines() but inserts a break before the first character that would cause the line to exceed the given max_width.

fn lines_wrapped_by_character<'a>(&'a self, font_size: FontSize, text: &'a str, max_width: Scalar) -> LinesWrappedByCharacter<'a, C> where C: CharacterCache

An iterator that behaves the same as text.lines() but inserts a break before the first character that would cause the line to exceed the given max_width.

fn lines_wrapped_by_whitespace<'a>(&'a self, font_size: FontSize, text: &'a str, max_width: Scalar) -> LinesWrappedByWhitespace<'a, C> where C: CharacterCache

An iterator that behaves the same as text.lines() but inserts a break before the first character that would cause the line to exceed the given max_width.