Module conrod::graph::algo

This module was created in order to keep the graph module clean and focused upon the Graph data structure behaviour.

This module hosts more complex algorithms in which the Graph is a key component in producing the desired result.

Structs

PickScrollableWidgets
A node "walker" that yields all scrollable widgets under the given xy position in order from top to bottom.

PickWidgets
A node "walker" that yields all widgets under the given xy position in order from top to bottom.

Functions

cropped_area_of_widget
The rectangle that represents the maximum visible area for the widget with the given index.

cropped_area_of_widget_within_depth
The rectangle that represents the maximum visible area for the widget with the given index.

kids_bounding_box
Find the absolute Rect that bounds all widgets that are Depth children of the widget at the given idx.

pick_scrollable_widgets
Produces a graph node "walker" that yields all scrollable widgets under the given xy position in order from top to bottom.

pick_widgets
Produces a graph node "walker" that yields all widgets under the given xy position in order from top to bottom.

scroll_offset
Return the scroll_offset for the widget at the given index.