Function conrod::graph::algo::bounding_box

pub fn bounding_box<I: GraphIndex>(graph: &Graph, include_self: bool, maybe_target_xy: Option<Point>, use_kid_area: bool, idx: I, maybe_deepest_parent_idx: Option<NodeIndex>) -> Option<Rect>

The box that bounds the widget with the given ID as well as all its widget kids.

Bounds are given as (max y, min y, min x, max x) from the given target xy position.

If no target xy is given, the bounds will be given relative to the centre of the widget.

If use_kid_area is true, then bounds will be calculated relative to the centre of the kid_area of the widget, rather than the regular dimensions.

The maybe_deepest_parent_idx refers to the index of the parent that began the recursion, and is used when calculating the "visible area" for each widget. If None is given, it means we are the first widget in the recursion, and we will use ourselves as the deepest_parent_idx.