Function conrod::graph::algo::kids_bounding_box

pub fn kids_bounding_box(graph: &Graph,
                         prev_updated: &HashSet<Id>,
                         idx: Id)
                         -> Option<Rect>

Find the absolute Rect that bounds all widgets that are Depth children of the widget at the given idx.

FIXME: This currently uses call stack recursion to do a depth-first search through all depth_children for the total bounding box. This should use a proper Dfs type with it's own stack for safer traversal that won't blow the stack on hugely deep GUIs.