Struct conrod::graph::algo::PickScrollableWidgets

pub struct PickScrollableWidgets { /* fields omitted */ }

A node "walker" that yields all scrollable widgets under the given xy position in order from top to bottom.

Methods

impl PickScrollableWidgets

pub fn next(
&mut self,
graph: &Graph,
depth_order: &[Id],
theme: &Theme
) -> Option<Id>

The widget::Id of the next scrollable Widget under the xy location.

The Graph is traversed from the top down.