Struct conrod::image::Map

pub struct Map<Img> {
    pub trigger_redraw: Cell<bool>,
    // some fields omitted
}

A type used to map the widget::Id of Image widgets to their associated Img data.

The image::Map type is usually instantiated and loaded during the "setup" stage of the application before the main loop begins. A macro is provided to simplify the construction of maps with multiple images.

let image_map = image_map! {
    (RUST_LOGO, try!(image::open("rust-logo.png"))),
    (CAT_PIC, try!(image::open("floof.jpeg"))),
};

Fields
trigger_redraw: Cell<bool>

Whether or not the image::Map will trigger a redraw the next time Ui::draw is called.

This is automatically set to true when any method that takes &mut self is called.

Methods

impl<Img> Map<Img>

fn new() -> Self

Construct a new, empty image::Map.

fn get_mut(&mut self, id: Id) -> Option<&mut Img>

Uniquely borrow the Img associated with the given widget.

Note: Calling this will trigger a redraw the next time Ui::draw_if_changed is called.

fn insert(&mut self, img: Img) -> Id

Inserts the given image into the map, returning its associated image::Id. The user must store the returned image::Id in order to use, modify or remove the inserted image.

Note: Calling this will trigger a redraw the next time Ui::draw_if_changed is called.

fn replace(&mut self, id: Id, img: Img) -> Option<Img>

Replaces the given image in the map if it exists. Returns the image or None.

Note: Calling this will trigger a redraw the next time Ui::draw_if_changed is called.

fn remove(&mut self, id: Id) -> Option<Img>

Removes the given image from the map if it exists. Returns the image or None.

Any future use of the given image::Id will be invalid.

Note: Calling this will trigger a redraw the next time Ui::draw_if_changed is called.

fn extend<I>(&mut self, images: I) -> NewIds where I: IntoIterator<Item=Img>

Insert each of the images yielded by the given iterator and produce an iterator yielding their generated Ids in the same order.

Note: Calling this will trigger a redraw the next time Ui::draw_if_changed is called.