Module conrod::input::global

Handles all of the global input events and state. The core of this module is the Global struct. It is responsible for aggregating and interpreting raw input events into high-level semantic events.

Structs

Events
Iterator over all global event::Events that have occurred since the last time Ui::set_widgets was called.

Global
Global input event handler that also implements input::Provider. The Ui passes all events to it's Global instance, which aggregates and interprets the events to provide so-called 'high-level' events to widgets. This input gets reset after every update by the Ui.

UiEvents
An iterator yielding all event::Uis that have occurred since the last time Ui::set_widgets was called.