Enum conrod::input::state::mouse::ButtonPosition

pub enum ButtonPosition {
    Up,
    Down(Point, Option<Id>),
}

Whether the button is up or down.

Variants
Up

The button is up (i.e. pressed).

Down(Point, Option<Id>)

The button is down and was originally pressed down at the given Point over the widget at the given widget::Id.

Methods

impl ButtonPosition

fn relative_to(self, xy: Point) -> Self

If the mouse button is down, return a new one with position relative to the given xy.

fn is_down(&self) -> bool

Is the ButtonPosition down.

fn is_up(&self) -> bool

Is the ButtonPosition up.

fn if_down(&self) -> Option<(Point, Option<Id>)>

Returns the position at which the button was pressed along with the widget that was under the mouse at the time of pressing if the position is Down.

fn xy_if_down(&self) -> Option<Point>

Returns the position at which the button was pressed if it's down.