Struct conrod::input::state::mouse::ButtonMap

pub struct ButtonMap { /* fields omitted */ }

Stores the state of all mouse buttons.

If the mouse button is down, it stores the position of the mouse when the button was pressed

Methods

impl ButtonMap

fn new() -> Self

Returns a new button map with all states set to None

fn relative_to(self, xy: Point) -> Self

Returns a copy of the ButtonMap relative to the given Point

fn left(&self) -> &ButtonPosition

The state of the left mouse button.

fn middle(&self) -> &ButtonPosition

The state of the middle mouse button.

fn right(&self) -> &ButtonPosition

The state of the right mouse button.

fn press(&mut self, button: Button, xy: Point, widget: Option<Id>)

Sets the Button in the Down position.

fn release(&mut self, button: Button)

Set's the Button in the Up position.

fn pressed(&self) -> PressedButtons

An iterator yielding all pressed mouse buttons along with the location at which they were originally pressed.