Struct conrod::input::state::State

pub struct State {
    pub mouse: Mouse,
    pub widget_capturing_keyboard: Option<Id>,
    pub widget_capturing_mouse: Option<Id>,
    pub widget_under_mouse: Option<Id>,
    pub modifiers: ModifierKey,
}

Holds the current state of user input.

This includes the state of all buttons on the keyboard and mouse, as well as the position of the mouse.

It also includes which widgets, if any, are capturing keyboard and mouse input.

Fields
mouse: Mouse

Mouse position and button state.

widget_capturing_keyboard: Option<Id>

Which widget, if any, is currently capturing the keyboard

widget_capturing_mouse: Option<Id>

Which widget, if any, is currently capturing the mouse

widget_under_mouse: Option<Id>

The widget that is currently under the mouse cursor.

If the mouse is currently over multiple widgets, this index will represent the top-most, non-graphic-child widget.

modifiers: ModifierKey

Which modifier keys are being held down.

Methods

impl State

fn new() -> State

Returns a fresh new input state

fn relative_to(self, xy: Point) -> State

Returns a copy of the input::State relative to the given position::Point