Module conrod::input::widget

Contains all the logic for filtering input events and making them relative to widgets.

The core of this module is the Widget::for_widget method, which creates an InputProvider that provides input events for a specific widget.

Structs

ButtonClicks
An Iterator yielding all mouse button clicks occuring within the given sequence of widget::Clicks.

ButtonDrags
An Iterator yielding all mouse button drags occuring within the given sequence of widget::Drags.

Clicks
An Iterator yielding all mouse clicks occuring within the given sequence of widget::Events.

Drags
An iterator that yields all event::Drag events yielded by the Events iterator.

Events
An iterator yielding all events that are relevant to a specific widget.

KeyPresses
An Iterator yielding all keyboard button presses occuring within the given sequence of Presses.

KeyReleases
An Iterator yielding all keyboard button releases occuring within the given sequence of Releases.

Mouse
A view of the input::state::Mouse that is specific to a single widget.

MouseButtonPresses
An Iterator yielding all mouse button presses occuring within the given sequence of Presses for some specific mouse button.

MouseButtonReleases
An Iterator yielding all mouse button releases occuring within the given sequence of Releases for some specific mouse button.

MousePresses
An Iterator yielding all mouse button presses occuring within the given sequence of Presses.

MouseReleases
An Iterator yielding all mouse button releases occuring within the given sequence of Releases.

Presses
An Iterator yielding all button presses occuring within the given sequence of widget::Events.

Releases
An Iterator yielding all button releases occuring within the given sequence of widget::Events.

Scrolls
An iterator that yields all Scroll events yielded by the given Events iterator.

Texts
An iterator that yields all Input::Text events yielded by the Events iterator.

Widget
Provides only events and input state that are relevant to a specific widget.