Struct conrod::input::widget::Mouse

pub struct Mouse<'a> {
    pub buttons: &'a ButtonMap,
    // some fields omitted
}

A view of the input::state::Mouse that is specific to a single widget.

Fields
buttons: &'a ButtonMap

The state of each MouseButton.

Methods

impl<'a> Mouse<'a>

fn abs_xy(&self) -> Point

The absolute position of the mouse within the window.

fn rel_xy(&self) -> Point

The position of the mouse relative to the middle of the widget's Rect.

fn is_over(&self) -> bool

Is the mouse currently over the widget.