Module conrod::position

Items related to 2D positioning, used throughout conrod.

Reexports

pub use self::range::Edge;

pub use self::range::Range;

pub use self::rect::Corner;

pub use self::rect::Rect;

Modules

range
A type for working one-dimensional ranges.

rect
Defines the scalar Rect type used throughout conrod.

Structs

Padding
The distance between the inner edge of a border and the outer edge of the inner content.

Enums

Align
The orientation of Alignment along some Axis.

Axis
Represents either Axis in the 2-dimensional plane.

Dimension
The length of a Widget over either the x or y axes.

Direction
Directionally positioned, normally relative to some other widget.

Place
Place the widget at a position on some other widget.

Position
Some Position of some Widget along a single axis.

Relative
Positions that are described as Relative to some other Widget.

Traits

Positionable
Widgets that are positionable.

Sizeable
Widgets that support different dimensions.

Type Definitions

Depth
The depth at which the widget will be rendered.

Dimensions
General use 2D spatial dimensions.

Margin
The margin for some Placement on either end of an axis.

Point
General use 2D spatial point.

Scalar
An alias over the Scalar type used throughout Conrod.