Enum conrod::position::Dimension

pub enum Dimension {
    Absolute(Scalar),
    Of(Id, Option<Scalar>),
    KidAreaOf(Id, Option<Scalar>),
}

The length of a Widget over either the x or y axes.

This type is used to represent the different ways in which a dimension may be sized.

See the Sizeable trait for methods that allow for setting the x and y Dimensions in various ways.

Note that Sizeable is implemented for all types that implement Widget.

Variants
Absolute(Scalar)

Some specific length has been given.

Of(Id, Option<Scalar>)

The dimension should match that of the widget at the given index.

The Option<Scalar> is an optional padding argument which when Some, will subtract the scalar from both ends of the other widget's dimension.

KidAreaOf(Id, Option<Scalar>)

The dimension should match that of the kid_area of the widget at the given index.

The Option<Scalar> is an optional padding argument which when Some, will subtract the scalar from both ends of the other widget's dimension.