Enum conrod::position::Relative

pub enum Relative {
    Scalar(Scalar),
    Align(Align),
    Direction(Direction, Scalar),
    Place(Place),
}

Positions that are described as Relative to some other Widget.

Relative describes a relative position along a single axis.

Variants
Scalar(Scalar)

A relative scalar distance.

Align(Align)

Aligned to either the Start, Middle or End.

Direction(Direction, Scalar)

A distance as a Scalar value over the given Direction.

Place(Place)

Some place on top of another widget.

Similar to Align, but represents the Start/End of the other widget's kid_area.

Also allows for specifying a Margin from either end.

Using Place allows the Ui to infer the widget's parent as the widget upon which it is Placed, though this inferrence only occurs if the parent was not specifically set.