Module conrod::render

Conrod's generic graphics backend.

Note: Conrod currently uses Piston's generic graphics crate (and specifically the Graphics) trait to enable genericity over custom user backends. This dependency may change in the near future in favour of a simplified conrod-specific graphics and character caching backend trait.

This is the only module in which the piston graphics crate will be used directly.

Structs

OwnedPrimitives
An owned alternative to the Primitives type.

Primitive
Data required for rendering a single primitive widget.

Primitives
An iterator-like type that yields a reference to each primitive in order of depth for rendering.

Text
A type used for producing a PositionedGlyph iterator.

WalkOwnedPrimitives
An iterator-like type for yielding Primitives from an OwnedPrimitives.

Enums

PrimitiveKind
The unique kind for each primitive element in the Ui.

Traits

PrimitiveWalker
A trait that allows the user to remain generic over types yielding Primitives.