Struct conrod::render::Primitive

pub struct Primitive<'a> {
    pub id: Id,
    pub kind: PrimitiveKind<'a>,
    pub scizzor: Rect,
    pub rect: Rect,
}

Data required for rendering a single primitive widget.

Fields
id: Id

The id of the widget within the widget graph.

kind: PrimitiveKind<'a>

State and style for this primitive widget.

scizzor: Rect

The Rect to which the primitive widget should be cropped.

Only parts of the widget within this Rect should be drawn.

rect: Rect

The bounding rectangle for the Primitive.