Trait conrod::render::PrimitiveWalker

pub trait PrimitiveWalker {
    fn next_primitive(&mut self) -> Option<Primitive>;
}

A trait that allows the user to remain generic over types yielding Primitives.

This trait is implemented for both the Primitives and WalkOwnedPrimitives types.

Required Methods

fn next_primitive(&mut self) -> Option<Primitive>

Yield the next Primitive in order of depth, bottom to top.