Struct conrod::scroll::Bar

pub struct Bar {
    pub interaction: Interaction,
    pub offset: Scalar,
    pub max_offset: Scalar,
    pub total_length: Scalar,
}

The state of a single scrollbar.

Fields

interaction

The current interaction with the Scrollbar.

offset

The current scroll position as an offset from the top left.

max_offset

The maximum possible offset for the handle.

total_length

The total length of the area occupied by child widgets.