Struct conrod::scroll::Style

pub struct Style {
    pub maybe_thickness: Option<Scalar>,
    pub maybe_color: Option<Color>,
}

Style for the Scrolling.

Fields

maybe_thickness

The width for vertical scrollbars, the height for horizontal scrollbars.

maybe_color

The color of the scrollbar.

Methods

impl Style

fn new() -> Style

Construct a new default Style.

fn thickness(&self, theme: &Theme) -> Scalar

Get the thickness of the scrollbar or a default from the theme.

fn color(&self, theme: &Theme) -> Color

Get the Color for the scrollbar.

impl Style

fn new() -> Style

Construct a new default Style.

fn thickness(&self, theme: &Theme) -> Scalar

Get the thickness of the scrollbar or a default from the theme.

fn color(&self, theme: &Theme) -> Color

Get the Color for the scrollbar.