Trait conrod::scroll::Axis

pub trait Axis {
    fn parallel_range(Rect) -> Range;
    fn perpendicular_range(Rect) -> Range;
    fn padding_range(Padding) -> Range;
    fn track(container: Rect, thickness: Scalar) -> Rect;
    fn mouse_scalar(mouse_xy: Point) -> Scalar;
    fn mouse_scroll_axis(MouseScroll) -> Scalar;
    fn handle_rect(perpendicular_track_range: Range, handle_range: Range) -> Rect;
    fn offset_direction() -> Scalar;
}

Methods for distinguishing behaviour between both scroll axes at compile-time.

Required Methods

fn parallel_range(Rect) -> Range

The range of the given Rect that is parallel with this Axis.

fn perpendicular_range(Rect) -> Range

The range of the given Rect that is perpendicular with this Axis.

fn padding_range(Padding) -> Range

Given some rectangular Padding, return the Range that corresponds with this Axis.

fn track(container: Rect, thickness: Scalar) -> Rect

The Rect for a scroll "track" with the given thickness for a container with the given Rect.

fn mouse_scalar(mouse_xy: Point) -> Scalar

The coordinate of the given mouse position that corresponds with this Axis.

fn mouse_scroll_axis(MouseScroll) -> Scalar

The coordinate of the given MouseScroll that corresponds with this Axis.

fn handle_rect(perpendicular_track_range: Range, handle_range: Range) -> Rect

The Rect for a scroll handle given both Ranges.

fn offset_direction() -> Scalar

A Scalar multiplier representing the direction in which positive offset shifts the scrollable_range (either -1.0 or `1.0).