Struct conrod::Canvas

pub struct Canvas<'a> {
    pub common: CommonBuilder,
    pub style: Style,
    pub maybe_title_bar_label: Option<&'a str>,
    pub maybe_splits: Option<FlowOfSplits<'a>>,
}

Canvas is designed to be a "container"-like "parent" widget that simplifies placement of "children" widgets.

Widgets can be placed on a Canvas in a variety of ways using methods from the Positionable trait.

Canvas provides methods for padding the kid widget area which can make using the Place-related Positionable methods a little easier.

A Canvas can also be divided into a sequence of smaller Canvasses using the .flow_* methods. This creates a kind of Canvas tree, where each "split" can be sized using the .length or .length_weight methods.

See the canvas.rs example for a demonstration of the Canvas type.

Fields

common

Data necessary and common for all widget builder types.

style

The builder data related to the style of the Canvas.

maybe_title_bar_label

The label for the Canvas' TitleBar if there is one.

maybe_splits

A list of child Canvasses as splits of this Canvas flowing in the given direction.

Methods

impl<'a> Canvas<'a>

fn new() -> Self

Construct a new Canvas builder.

fn title_bar(self, title_bar: &'a str) -> Self

Build the type's self.$($assignee).+ with the given $Type.

fn pad_left(self, pad_left: Scalar) -> Self

Build the type's self.$($assignee).+ with the given $Type.

fn pad_right(self, pad_right: Scalar) -> Self

Build the type's self.$($assignee).+ with the given $Type.

fn pad_bottom(self, pad_bottom: Scalar) -> Self

Build the type's self.$($assignee).+ with the given $Type.

fn pad_top(self, pad_top: Scalar) -> Self

Build the type's self.$($assignee).+ with the given $Type.

fn with_style(self, with_style: Style) -> Self

Build the type's self.$($assignee).+ with the given $Type.

fn length(self, length: Scalar) -> Self

Set the length of the Split as an absolute scalar.

fn length_weight(self, weight: Weight) -> Self

Set the length of the Split as a weight.

The default length weight for each widget is 1.0.

fn flow_right(self, splits: ListOfSplits<'a>) -> Self

Set the child Canvasses flowing to the right.

fn flow_left(self, splits: ListOfSplits<'a>) -> Self

Set the child Canvasses flowing to the left.

fn flow_up(self, splits: ListOfSplits<'a>) -> Self

Set the child Canvasses flowing upwards.

fn flow_down(self, splits: ListOfSplits<'a>) -> Self

Set the child Canvasses flowing downwards.

fn pad(self, pad: Scalar) -> Self

Set the padding for all edges of the area where child widgets will be placed.

fn padding(self, pad: Padding) -> Self

Set the padding of the area where child widgets will be placed.