Struct conrod::CommonBuilder

pub struct CommonBuilder {
    pub maybe_width: Option<Scalar>,
    pub maybe_height: Option<Scalar>,
    pub maybe_position: Option<Position>,
    pub maybe_h_align: Option<HorizontalAlign>,
    pub maybe_v_align: Option<VerticalAlign>,
    pub maybe_depth: Option<Depth>,
    pub maybe_parent_id: MaybeParent,
    pub is_floating: bool,
}

A struct containing builder data common to all Widget types. This type allows us to do a blanket impl of Positionable and Sizeable for T: Widget.

Fields

maybe_width

The width of a Widget.

maybe_height

The height of a Widget.

maybe_position

The position of a Widget.

maybe_h_align

The horizontal alignment of a Widget.

maybe_v_align

The vertical alignment of a Widget.

maybe_depth

The rendering Depth of the Widget.

maybe_parent_id

The parent widget of the Widget.

is_floating

Whether or not the Widget is a "floating" Widget.

Methods

impl CommonBuilder

fn new() -> CommonBuilder

Construct an empty, initialised CommonBuilder.