Struct conrod::EnvelopeEditorStyle

pub struct EnvelopeEditorStyle {
    pub color: Option<Color>,
    pub frame: Option<f64>,
    pub frame_color: Option<Color>,
    pub label_color: Option<Color>,
    pub label_font_size: Option<FontSize>,
    pub value_font_size: Option<FontSize>,
    pub point_radius: Option<Scalar>,
    pub line_thickness: Option<Scalar>,
}

Styling for the EnvelopeEditor, necessary for constructing its renderable Element.

Fields

color

Coloring for the EnvelopeEditor's FramedRectangle.

frame

Thickness of the FramedRectangle's frame.

frame_color

Color of the frame.

label_color

Color of the label.

label_font_size

The font size of the EnvelopeEditor's label if one was given.

value_font_size

The font size of the value label.

point_radius

The radius of the envelope points.

line_thickness

The thickness of the envelope lines.

Methods

impl Style

fn new() -> Self

Construct the default Style, initialising all fields to None.

impl Style

fn color(&self, theme: &Theme) -> Color

Retrieves the value from the Style.

If the Style's field is None, falls back to default specified within the Theme.

fn frame(&self, theme: &Theme) -> f64

Retrieves the value from the Style.

If the Style's field is None, falls back to default specified within the Theme.

fn frame_color(&self, theme: &Theme) -> Color

Retrieves the value from the Style.

If the Style's field is None, falls back to default specified within the Theme.

fn label_color(&self, theme: &Theme) -> Color

Retrieves the value from the Style.

If the Style's field is None, falls back to default specified within the Theme.

fn label_font_size(&self, theme: &Theme) -> FontSize

Retrieves the value from the Style.

If the Style's field is None, falls back to default specified within the Theme.

fn value_font_size(&self, theme: &Theme) -> FontSize

Retrieves the value from the Style.

If the Style's field is None, falls back to default specified within the Theme.

fn point_radius(&self, theme: &Theme) -> Scalar

Retrieves the value from the Style.

If the Style's field is None, falls back to default specified within the Theme.

fn line_thickness(&self, theme: &Theme) -> Scalar

Retrieves the value from the Style.

If the Style's field is None, falls back to default specified within the Theme.