Struct conrod::FramedRectangle

pub struct FramedRectangle {
    pub common: CommonBuilder,
    pub style: Style,
}

A filled rectangle widget that may or may not have some frame.

Fields

common

Data necessary and common for all widget builder types.

style

Unique styling for the FramedRectangle.

Methods

impl FramedRectangle

fn new(dim: Dimensions) -> Self

Build a new FramedRectangle.

fn with_style(self, with_style: Style) -> Self

Build the type's self.$($assignee).+ with the given $Type.