Struct conrod::Label

pub struct Label<'a> {
    // some fields omitted
}

Displays some given text centred within a rectangle.

Methods

impl<'a> Label<'a>

fn new(text: &'a str) -> Label<'a>

Construct a new Label widget.

fn size(self, size: FontSize) -> Label<'a>

Set the font size for the label.

fn set<C>(self, ui_id: UiId, ui: &mut Ui<C>) where C: CharacterCache

After building the Label, use this method to set its current state into the given Ui. It will use this state for rendering the next time ui.draw(graphics) is called.