Struct conrod::MouseButtonState

pub struct MouseButtonState {
    pub was_just_pressed: bool,
    pub was_just_released: bool,
    pub position: ButtonPosition,
}

The current state of a Mouse button.

Fields

was_just_pressed

The button has been pressed since the last update.

was_just_released

The button has been released since the last update.

position

The current position of the button.

Methods

impl ButtonState

fn new() -> ButtonState

Constructor for a default ButtonState.

fn reset_pressed_and_released(&mut self)

Reset the was_just_released and was_just_pressed flags.