Struct conrod::NumberDialer

pub struct NumberDialer<'a, T, F> {
    // some fields omitted
}

A widget for precision control over any digit within a value. The reaction is triggered when the value is updated or if the mouse button is released while the cursor is above the widget.

Methods

impl<'a, T: Float, F> NumberDialer<'a, T, F>

fn new(value: T, min: T, max: T, precision: u8) -> NumberDialer<'a, T, F>

Construct a new NumberDialer widget.

fn react(self, reaction: F) -> NumberDialer<'a, T, F>

Set the reaction for the NumberDialer. It will be triggered when the value is updated or if the mouse button is released while the cursor is above the widget.

fn enabled(self, flag: bool) -> Self

If true, will allow user inputs. If false, will disallow user inputs.