Struct conrod::PointPath

pub struct PointPath<I> {
    pub points: I,
    pub common: CommonBuilder,
    pub style: Style,
    pub maybe_shift_to_centre_from: Option<Point>,
}

A simple, non-interactive widget for drawing a series of lines and/or points.

Fields

points

Some iterator yielding a series of Points.

common

Data necessary and common for all widget builder types.

style

Unique styling for the PointPath.

maybe_shift_to_centre_from

Whether or not the points should be automatically centred to the widget position.

Methods

impl<I> PointPath<I>

fn styled(points: I, style: Style) -> Self

The same as PointPath::new but with th given style.

fn new(points: I) -> Self

Build a new default PointPath widget.

It is recommended that you also see the abs and centred constructors for smart positioning and layout.

fn abs(points: I) -> Self where I: IntoIterator<Item=Point> + Clone

Build a new PointPath whose bounding box is fit to the absolute co-ordinates of the points.

This requires that the points iterator is Clone so that we may iterate through and determine the bounding box of the points.

If you would rather centre the points to the middle of the bounding box, use PointPath::centred instead.

fn abs_styled(points: I, style: Style) -> Self where I: IntoIterator<Item=Point> + Clone

The same as PointPath::abs but constructs the PointPath with the given style.

fn centred(points: I) -> Self where I: IntoIterator<Item=Point> + Clone

Build a new PointPath and shift the location of the points so that the centre of their bounding rectangle lies at the position determined for the PointPath widget.

This is useful if your points simply describe a shape and you want to position them using conrod's auto-layout or Positionable methods.

If you would rather centre the bounding box to the points, use PointPath::abs instead.

fn centred_styled(points: I, style: Style) -> Self where I: IntoIterator<Item=Point> + Clone

The same as PointPath::centred but constructs the PointPath with the given style.

fn thickness(self, thickness: Scalar) -> Self

The thickness or width of the PointPath's lines.

Use this instead of Positionable::width for the thickness of the Line, as width and height refer to the dimensions of the bounding rectangle.

fn solid(self) -> Self

Make a Solid line.

fn dashed(self) -> Self

Make a line with a Dashed pattern.

fn dotted(self) -> Self

Make a line with a Dotted pattern.