Struct conrod::PositionMatrix

pub struct PositionMatrix {
    // some fields omitted
}

A type to simplify placement of various widgets in a matrix or grid layout.

Methods

impl Matrix

fn new(cols: usize, rows: usize) -> Matrix

Start building a new position Matrix.

fn cell_padding(self, w: Scalar, h: Scalar) -> Matrix

Produce the matrix with the given cell padding.

fn each_widget<C, F>(self, ui: &mut Ui<C>, f: F) where C: CharacterCache, F: FnMut(&mut Ui<C>, WidgetNum, ColNum, RowNum, Point, Dimensions)

Call the given function for every element in the Matrix.