Struct conrod::Slider

pub struct Slider<'a, T, F> {
    pub skew: f32,
    pub maybe_react: Option<F>,
    pub enabled: bool,
    // some fields omitted
}

Linear value selection. If the slider's width is greater than it's height, it will automatically become a horizontal slider, otherwise it will be a vertical slider. Its reaction is triggered if the value is updated or if the mouse button is released while the cursor is above the rectangle.

Fields

skew

The amount in which the slider's display should be skewed.

Higher skew amounts (above 1.0) will weight lower values.

Lower skew amounts (below 1.0) will weight heigher values.

All skew amounts should be greater than 0.0.

maybe_react

Set the reaction for the Slider.

It will be triggered if the value is updated or if the mouse button is released while the cursor is above the rectangle.

enabled

Whether or not user input is enabled for the Slider.

Methods

impl<'a, T, F> Slider<'a, T, F>

fn new(value: T, min: T, max: T) -> Self

Construct a new Slider widget.

fn skew(self, skew: f32) -> Self

Build the type's self.$($assignee).+ with the given $Type.

fn react(self, react: F) -> Self

Build the type's self.$($assignee).+ with the given $Type.

fn enabled(self, enabled: bool) -> Self

Build the type's self.$($assignee).+ with the given $Type.