Struct conrod::Split

pub struct Split<'a> {
    // some fields omitted
}

A type of Canvas for flexibly designing and guiding widget layout as splits of a window.

Methods

impl<'a> Split<'a>

fn new(id: CanvasId) -> Split<'a>

Construct a default Canvas Split.

fn length(self, length: Length) -> Split<'a>

Set the dimension of the Split.

fn flow(self, dir: Direction, splits: &'a [Split<'a>]) -> Split<'a>

Set the child Canvas Splits of the current Canvas flowing in a given direction.

fn flow_down(self, splits: &'a [Split<'a>]) -> Split<'a>

Set the child Canvasses flowing downwards.

fn flow_up(self, splits: &'a [Split<'a>]) -> Split<'a>

Set the child Canvasses flowing upwards.

fn flow_right(self, splits: &'a [Split<'a>]) -> Split<'a>

Set the child Canvasses flowing to the right.

fn flow_left(self, splits: &'a [Split<'a>]) -> Split<'a>

Set the child Canvasses flowing to the left.

fn pad_left(self, pad: Scalar) -> Split<'a>

Set the padding from the left edge.

fn pad_right(self, pad: Scalar) -> Split<'a>

Set the padding from the right edge.

fn pad_top(self, pad: Scalar) -> Split<'a>

Set the padding from the top edge.

fn pad_bottom(self, pad: Scalar) -> Split<'a>

Set the padding from the bottom edge.

fn pad(self, pad: Scalar) -> Split<'a>

Set the padding for all edges.

fn margin_left(self, pad: Scalar) -> Split<'a>

Set the margin from the left edge.

fn margin_right(self, pad: Scalar) -> Split<'a>

Set the margin from the right edge.

fn margin_top(self, pad: Scalar) -> Split<'a>

Set the margin from the top edge.

fn margin_bottom(self, pad: Scalar) -> Split<'a>

Set the margin from the bottom edge.

fn margin(self, pad: Scalar) -> Split<'a>

Set the margin for all edges.

fn set<C>(self, ui: &mut Ui<C>)

Store the Canvas and it's children within the Ui. Each Canvas can be accessed via it's unique identifier CanvasId.