Struct conrod::Tabs

pub struct Tabs<'a> {
    // some fields omitted
}

A wrapper around a list of canvasses that displays thema s a list of selectable tabs.

Methods

impl<'a> Tabs<'a>

fn new(tabs: &'a [(Id, &'a str)]) -> Tabs<'a>

Construct some new Canvas Tabs.

fn starting_canvas(self, canvas_id: Id) -> Self

Set the initially selected tab with a Canvas via its widget::Id.

fn pad(self, pad: Scalar) -> Tabs<'a>

Set the padding for all edges.

fn layout_horizontally(self) -> Self

Layout the tabs horizontally.

fn layout_vertically(self) -> Self

Layout the tabs vertically.

fn bar_width(self, bar_width: Scalar) -> Self

Build the type's self.$($assignee).+ with the given $Type.

fn starting_tab_idx(self, starting_tab_idx: usize) -> Self

Build the type's self.$($assignee).+ with the given $Type.

fn label_color(self, label_color: Color) -> Self

Build the type's self.$($assignee).+ with the given $Type.

fn label_font_size(self, label_font_size: FontSize) -> Self

Build the type's self.$($assignee).+ with the given $Type.

fn canvas_style(self, canvas_style: Style) -> Self

Build the type's self.$($assignee).+ with the given $Type.

fn pad_left(self, pad_left: Scalar) -> Self

Build the type's self.$($assignee).+ with the given $Type.

fn pad_right(self, pad_right: Scalar) -> Self

Build the type's self.$($assignee).+ with the given $Type.

fn pad_bottom(self, pad_bottom: Scalar) -> Self

Build the type's self.$($assignee).+ with the given $Type.

fn pad_top(self, pad_top: Scalar) -> Self

Build the type's self.$($assignee).+ with the given $Type.