Struct conrod::TabsStyle

pub struct TabsStyle {
    pub maybe_layout: Option<Layout>,
    pub maybe_bar_width: Option<Scalar>,
    pub maybe_label_color: Option<Color>,
    pub maybe_label_font_size: Option<FontSize>,
    pub canvas: Style,
}

The styling for Canvas Tabs.

Fields

maybe_layout

The direction in which the tabs will be laid out.

maybe_bar_width

The width of the tab bar. For horizontally laid out tabs, this is the height of the bar. For vertically laid out tabs, this is the width of the bar.

maybe_label_color

The color of the tabs' labels.

maybe_label_font_size

The font size for the tabs' labels.

canvas

Styling for each of the canvasses passed to the Canvas.

Methods

impl Style

fn new() -> Style

Construct the default Tabs style.

fn layout(&self, theme: &Theme) -> Layout

Get the layout of the tabs for the Tabs widget.

fn label_color(&self, theme: &Theme) -> Color

Get the color for the tab labels.

fn font_size(&self, theme: &Theme) -> FontSize

Get the font size for the tab labels.