Struct conrod::Text

pub struct Text<'a> {
    pub common: CommonBuilder,
    pub text: &'a str,
    pub style: Style,
}

Displays some given text centred within a rectangular area.

By default, the rectangular dimensions are fit to the area occuppied by the text.

If some horizontal dimension is given, the text will automatically wrap to the width and align in accordance with the produced Align.

Fields

common

Data necessary and common for all widget builder types.

text

The text to be drawn by the Text.

style

Unique styling for the Text.

Methods

impl<'a> Text<'a>

fn new(text: &'a str) -> Self

Build a new Text widget.

fn font_size(self, size: FontSize) -> Self

Build the Text with the given font size.

fn no_line_wrap(self) -> Self

Specify that the Text should not wrap lines around the width.

fn wrap_by_word(self) -> Self

Line wrap the Text at the beginning of the first word that exceeds the width.

fn wrap_by_character(self) -> Self

Line wrap the Text at the beginning of the first character that exceeds the width.

fn with_style(self, style: Style) -> Self

Build the Text with the given Style.

fn align_text_left(self) -> Self

Align the text to the left of its bounding Rect's x axis range.

fn align_text_middle(self) -> Self

Align the text to the middle of its bounding Rect's x axis range.

fn align_text_right(self) -> Self

Align the text to the right of its bounding Rect's x axis range.

fn line_spacing(self, height: Scalar) -> Self

The height of the space used between consecutive lines.