Struct conrod::TextBox

pub struct TextBox<'a, F> {
    // some fields omitted
}

A widget for displaying and mutating a given one-line text String. It's reaction is triggered upon pressing of the Enter/Return key.

Methods

impl<'a, F> TextBox<'a, F>

fn new(text: &'a mut String) -> TextBox<'a, F>

Construct a TextBox widget.

fn font_size(self, font_size: FontSize) -> TextBox<'a, F>

Set the font size of the text.

fn react(self, reaction: F) -> TextBox<'a, F>

Set the reaction for the TextBox. It will be triggered upon pressing of the Enter/Return key.

fn enabled(self, flag: bool) -> Self

If true, will allow user inputs. If false, will disallow user inputs.