Struct conrod::TextBoxStyle

pub struct TextBoxStyle {
    pub text_padding: Option<Scalar>,
    pub color: Option<Color>,
    pub frame: Option<Scalar>,
    pub frame_color: Option<Color>,
    pub text_color: Option<Color>,
    pub font_size: Option<FontSize>,
    pub x_align: Option<Align>,
}

Unique graphical styling for the TextBox.

Fields
text_padding: Option<Scalar>

The length of the gap between the bounding rectangle's frame and the edge of the text.

color: Option<Color>

Color of the rectangle behind the text.

If you don't want to see the rectangle, either set the color with a zeroed alpha or use the TextEdit widget directly.

frame: Option<Scalar>

The width of the bounding FramedRectangle frame.

frame_color: Option<Color>

The color of the FramedRecangle's frame.

text_color: Option<Color>

The color of the TextEdit widget.

font_size: Option<FontSize>

The font size for the text.

x_align: Option<Align>

The horizontal alignment of the text.

Methods

impl Style

fn new() -> Self

Construct the default Style, initialising all fields to None.

impl Style

fn text_padding(&self, theme: &Theme) -> Scalar

Retrieves the value from the Style.

If the Style's field is None, falls back to default specified within the Theme.

fn color(&self, theme: &Theme) -> Color

Retrieves the value from the Style.

If the Style's field is None, falls back to default specified within the Theme.

fn frame(&self, theme: &Theme) -> Scalar

Retrieves the value from the Style.

If the Style's field is None, falls back to default specified within the Theme.

fn frame_color(&self, theme: &Theme) -> Color

Retrieves the value from the Style.

If the Style's field is None, falls back to default specified within the Theme.

fn text_color(&self, theme: &Theme) -> Color

Retrieves the value from the Style.

If the Style's field is None, falls back to default specified within the Theme.

fn font_size(&self, theme: &Theme) -> FontSize

Retrieves the value from the Style.

If the Style's field is None, falls back to default specified within the Theme.

fn x_align(&self, theme: &Theme) -> Align

Retrieves the value from the Style.

If the Style's field is None, falls back to default specified within the Theme.