Struct conrod::TitleBar

pub struct TitleBar<'a, F> {
    pub common: CommonBuilder,
    pub style: Style,
    pub label: &'a str,
    pub maybe_react: Option<F>,
}

A simple title bar widget that automatically sizes itself to the top of some other widget.

Fields

common

Data necessary and common for all widget builder types.

style

Unique styling for the FramedRectangle.

label

A label displayed in the middle of the TitleBar.

maybe_react

Some function used to react to interactions with the TitleBar.

Methods

impl<'a, F> TitleBar<'a, F> where F: FnOnce(Interaction)

fn new<I>(label: &'a str, idx: I) -> Self where I: Into<Index> + Copy

Construct a new TitleBar widget and attach it to the widget at the given index.

fn label_style(self, style: TextStyle) -> Self

Pass some styling for the TitleBar's Label.

fn rect_style(self, style: FramedRectangleStyle) -> Self

Pass some styling for the TitleBar's FramedRectangle.

fn react(self, f: F) -> Self

Pass the title bar some function to call upon interaction changes.