Struct conrod::UiBuilder

pub struct UiBuilder {
    // some fields omitted
}

UiBuilder stores everything that needs to be customized when constructing

Methods

impl UiBuilder

fn new() -> Self

Create Ui builder with defaults.

maybe_theme: Theme::default()

maybe_widgets_capacity: None

fn theme(self, value: Theme) -> Self

Sets the theme of built Ui.

This method moves the current builder, so that it can be used in method chaining.

fn widgets_capacity(self, value: usize) -> Self

Sets widgets capacity of built Ui.

This method moves the current builder, so that it can be used in method chaining.

fn build(self) -> Ui

Build Ui from the given builder