Struct conrod::UpdateArgs

pub struct UpdateArgs<'a, W, C: 'a> where W: Widget {
    pub idx: Index,
    pub prev_state: &'a State<W::State>,
    pub rect: Rect,
    pub style: &'a W::Style,
    pub ui: UiCell<'a, C>,
}

Arguments for the Widget::update method in a struct to simplify the method signature.

Fields

idx

W's unique index.

prev_state

The Widget's state that was last returned by the update method.

rect

The rectangle describing the Widget's area.

style

The Widget's current Style.

ui

Restricted access to the Ui. Provides methods for immutably accessing the Ui's Theme and GlyphCache. Also allows calling Widget::set within the Widget::update method.