Struct conrod::WidgetState

pub struct WidgetState<'a, T: 'a> {
    // some fields omitted
}

A wrapper around a Widget's unique Widget::State.

This type is used to provide limited access to the Widget::State within the Widget::update method (to which it is passed via the UpdateArgs).

The type provides only two methods. One for viewing the state, the other for mutating it.

We do this so that we can keep track of whether or not the Widget::State has been mutated (using an internal has_updated flag). This allows us to know whether or not we need to produce a new Element for the Widget, without having to compare the previous and new Widget::States.

Methods

impl<'a, T> State<'a, T>

fn view(&self) -> &T

Immutably borrow the internal widget state.

fn update<F>(&mut self, f: F) where F: FnOnce(&mut T)

Mutate the internal widget state and set a flag notifying us that there has been a mutation.

If this method is not called, we assume that there has been no mutation and in turn we will re-use the Widget's pre-existing Element.

If this method is called, we assume that there has been some mutation and in turn will produce a new Element for the Widget. Thus, it is recommended that you only call this method if you need to update the unique state in some way.