Module conrod::text::font

The font::Id and font::Map types.

Structs

Id
A type-safe wrapper around the FontId.

Ids
Yields the Id for each Font within the Map.

Map
A collection of mappings from font::Ids to rusttype::Fonts.

NewIds
An iterator yielding an Id for each new rusttype::Font inserted into the Map via the insert_collection method.

Enums

Error
Returned when loading new fonts from file or bytes.

Functions

collection_from_file
Load a super::FontCollection from a file at a given path.

from_file
Load a single Font from a file at the given path.