Struct conrod::text::font::Map

pub struct Map { /* fields omitted */ }

A collection of mappings from font::Ids to rusttype::Fonts.

Methods

impl Map

fn new() -> Self

Construct the new, empty Map.

fn get(&self, id: Id) -> Option<&Font>

Borrow the rusttype::Font associated with the given font::Id.

fn insert(&mut self, font: Font) -> Id

Adds the given rusttype::Font to the Map and returns a unique Id for it.

fn insert_from_file<P>(&mut self, path: P) -> Result<Id, Error> where P: AsRef<Path>

Insert a single Font into the map by loading it from the given file path.

fn ids(&self) -> Ids

Produces an iterator yielding the Id for each Font within the Map.