Function conrod::text::glyph::index_after_cursor

pub fn index_after_cursor<I>(line_infos: I, cursor_idx: Index) -> Option<usize> where I: Iterator<Item=Info>

Find the index of the character that directly follows the cursor at the given cursor_idx.

Returns None if either the given cursor::Index line or idx fields are out of bounds of the line information yielded by the line_infos iterator.