Enum conrod::text::line::Break

pub enum Break {
    Wrap {
        byte: usize,
        char: usize,
        len_bytes: usize,
    },
    Newline {
        byte: usize,
        char: usize,
        len_bytes: usize,
    },
    End {
        byte: usize,
        char: usize,
    },
}

The two types of Break indices returned by the WrapIndicesBy iterators.

Variants
Wrap

A break caused by the text exceeding some maximum width.

Fields

byte: usize

The byte index at which the break occurs.

char: usize

The char index at which the string should wrap due to exceeding a maximum width.

len_bytes: usize

The byte length which should be skipped in order to reach the first non-whitespace character to use as the beginning of the next line.

Newline

A break caused by a newline character.

Fields

byte: usize

The byte index at which the string should wrap due to exceeding a maximum width.

char: usize

The char index at which the string should wrap due to exceeding a maximum width.

len_bytes: usize

The width of the "newline" token in bytes.

End

The end of the string has been reached, with the given length.

Fields

byte: usize

The ending byte index.

char: usize

The ending char index.

Methods

impl Break

fn byte_index(self) -> usize

Return the index at which the break occurs.

fn char_index(self) -> usize

Return the index of the char at which the break occurs.

To clarify, this index is to be used in relation to the Chars iterator.