Function conrod::text::line::infos

Important traits for Infos<'a, F>
impl<'a, F> Iterator for Infos<'a, F> where
F: for<'b> FnMut(&'b str, &'b Font, FontSize, Scalar) -> (Break, Scalar), type Item = Info;

pub fn infos<'a>(
    text: &'a str, 
    font: &'a Font, 
    font_size: FontSize
) -> Infos<'a, NextBreakFnPtr>

Produce an Infos iterator that yields an Info for every line in the given text.

The produced Infos iterator will not wrap the text, and only break each line via newline characters within the text (either \n or \r\n).