Function conrod::text::line::rects

pub fn rects<I>(infos: I,
                font_size: FontSize,
                bounding_rect: Rect,
                x_align: Justify,
                y_align: Align,
                line_spacing: Scalar)
                -> Rects<I> where I: Iterator<Item=Info> + ExactSizeIterator

Produce an iterator yielding the bounding Rect for each line in the text.

This function assumes that font_size is the same FontSize used to produce the Infos yielded by the infos Iterator.