Function conrod::text::line::selected_rects

pub fn selected_rects<'a, I>(lines_with_rects: I,
                             font: &'a Font,
                             font_size: FontSize,
                             start: Index,
                             end: Index)
                             -> SelectedRects<'a, I> where I: Iterator<Item=(&'a str, Rect)>

Produces an iterator yielding a Rect for the selected range in each selected line in a block of text.

The yielded Rects represent the selected range within each line of text.

Lines that do not contain any selected text will be skipped.