Module conrod::theme

Types a functionality for handling Canvas and Widget theming.

Structs

Theme
A serializable collection of canvas and widget styling defaults.

UniqueDefault
A WidgetDefault downcast to a Widget's unique Style type.

WidgetDefault
The defaults for a specific widget.