Enum conrod::utils::IterDiff

pub enum IterDiff<E, I> {
    FirstMismatch(usize, Chain<Once<E>, I>),
    Longer(Chain<Once<E>, I>),
    Shorter(usize),
}

A type returned by the iter_diff function.

Represents way in which the elements (of type E) yielded by the iterator I differ to some other iterator yielding borrowed elements of the same type.

I is some Iterator yielding elements of type E.

Variants
FirstMismatch(usize, Chain<Once<E>, I>)

The index of the first non-matching element along with the iterator's remaining elements starting with the first mis-matched element.

Longer(Chain<Once<E>, I>)

The remaining elements of the iterator.

Shorter(usize)

The total number of elements that were in the iterator.