Module conrod::widget

Widgets are the core building blocks for every conrod user interface.

This module contains items related to the implementation of the Widget trait. It also re-exports all widgets (and their modules) that are provided by conrod.

Reexports

pub use self::id::Id;

pub use self::primitive::line;

pub use self::primitive::line::Line;

pub use self::primitive::image;

pub use self::primitive::image::Image;

pub use self::primitive::point_path;

pub use self::primitive::point_path::PointPath;

pub use self::primitive::shape::circle;

pub use self::primitive::shape::circle::Circle;

pub use self::primitive::shape::oval;

pub use self::primitive::shape::oval::Oval;

pub use self::primitive::shape::polygon;

pub use self::primitive::shape::polygon::Polygon;

pub use self::primitive::shape::rectangle;

pub use self::primitive::shape::rectangle::Rectangle;

pub use self::primitive::text;

pub use self::primitive::text::Text;

pub use self::bordered_rectangle::BorderedRectangle;

pub use self::button::Button;

pub use self::canvas::Canvas;

pub use self::drop_down_list::DropDownList;

pub use self::list_select::ListSelect;

pub use self::envelope_editor::EnvelopeEditor;

pub use self::file_navigator::FileNavigator;

pub use self::list::List;

pub use self::matrix::Matrix;

pub use self::number_dialer::NumberDialer;

pub use self::plot_path::PlotPath;

pub use self::range_slider::RangeSlider;

pub use self::rounded_rectangle::RoundedRectangle;

pub use self::scrollbar::Scrollbar;

pub use self::slider::Slider;

pub use self::tabs::Tabs;

pub use self::text_box::TextBox;

pub use self::text_edit::TextEdit;

pub use self::title_bar::TitleBar;

pub use self::toggle::Toggle;

pub use self::xy_pad::XYPad;

Modules

bordered_rectangle
The BorderedRectangle widget and related items.

button
The Button widget and related items.

canvas
The Canvas widget and related items.

drop_down_list
The DropDownList and related items.

envelope_editor
The EnvelopeEditor widget and related items.

file_navigator
A widget for navigating through through a file system. Generally inspired by Finder.

id
The widget identifier type used throughout conrod, along with helper types and macros to simplify the process of generating them.

list
A helper widget, useful for instantiating a sequence of widgets in a vertical list.

list_select
A wrapper around the List widget providing the ability to select one or more items.

matrix
A helper widget for laying out child widgets in the form of a grid.

number_dialer
A widget for precision control over any base-10 digit within a given value.

plot_path
A widget for plotting a series of lines using the given function x -> y.

primitive
Primitive widgets are special in that they are built into conrod's rendering logic.

range_slider
A widget for specifying start and end values for some linear range.

rounded_rectangle
Similar to the Rectangle widget however is drawn with rounded corners.

scroll
Scroll related types and logic.

scrollbar
A widget that allows for manually scrolling via dragging the mouse.

slider
A widget for selecting a single value along some linear range.

tabs
A wrapper around a list of Canvasses that displays them as a list of selectable tabs.

text_box
A widget for displaying and mutating a one-line field of text.

text_edit
A widget for displaying and mutating multi-line text, given as a String.

title_bar
A simple title bar widget that automatically sizes itself to the top of some other widget.

toggle
A button that allows for toggling boolean state.

xy_pad
Used for displaying and controlling a 2D point on a cartesian plane within a given range.

Structs

CommonBuilder
A struct containing builder data common to all Widget types.

CommonState
A wrapper around state that is common to all Widget types.

CommonStyle
Styling and positioning data that is common between all widget types.

Floating
State necessary for "floating" (pop-up style) widgets.

KidArea
The area upon which a Widget's child widgets will be placed.

KidAreaArgs
Arguments to the Widget::kid_area method in a struct to simplify the method signature.

PostUpdateCache

PreUpdateCache

State
A wrapper around a Widget's unique Widget::State.

UpdateArgs
Arguments for the Widget::update method in a struct to simplify the method signature.

Enums

MaybeParent
The builder argument for the Widget's parent.

Traits

Style
The necessary bounds for a Widget's associated Style type.

Widget
A trait to be implemented by all Widget types.

Functions

default_x_dimension
Determines the default Dimension for a Widget.

default_y_dimension
Determines the default Dimension for a Widget.