Function conrod::widget::bordered_rectangle::rounded_border_triangles

Important traits for RoundedBorderTriangles
impl Iterator for RoundedBorderTriangles type Item = Triangle<Point>;

pub fn rounded_border_triangles(
    rect: Rect, 
    thickness: Scalar, 
    radius: Scalar, 
    corner_resolution: usize
) -> RoundedBorderTriangles

An iterator yielding triangles for a rounded border.

Clamps the thickness of the border to half the smallest dimension of the rectangle to ensure the bounds of the rect are not exceeded.