Struct conrod::widget::button::Button

pub struct Button<'a, S> {
    pub show: S,
    pub style: Style,
    // some fields omitted
}

A pressable button widget whose reaction is triggered upon release.

Fields
show: S

Whether the Button is a Flat color or an Image.

style: Style

Unique styling parameters for the Button.

Methods

impl<'a> Button<'a, Image>

pub fn image(image_id: Id) -> Self

Begin building a button displaying the given Image on top.

pub fn source_rectangle(self, rect: Rect) -> Self

The rectangular area of the image that we wish to display.

If this method is not called, the entire image will be used.

pub fn image_color(self, color: Color) -> Self

Map the Image's luminance to the given color.

pub fn image_color_with_feedback(self, color: Color) -> Self

Map the Image's luminance to the given color.

The color will change slightly when the button is highlighted or clicked to give the user some visual feedback.

pub fn hover_image(self, id: Id) -> Self

The image displayed while the mouse hovers over the Button.

pub fn press_image(self, id: Id) -> Self

The image displayed while the Button is pressed.

impl<'a> Button<'a, Flat>

pub fn new() -> Self

Begin building a flat-colored Button widget.

pub fn with_style(self, s: Style) -> Self

Override the default button style

pub fn hover_color(self, color: Color) -> Self

Specify a color to use when the mouse hovers over the button.

By default, this is color.highlighted() where color is the button's regular color.

pub fn press_color(self, color: Color) -> Self

Specify a color to use when the mouse presses the button.

By default, this is color.clicked() where color is the button's regular color.

impl<'a, S> Button<'a, S>

pub fn label_font_id(self, font_id: Id) -> Self

Specify the font used for displaying the label.

pub fn left_justify_label(self) -> Self

Align the label to the left of the Button's surface.

pub fn center_justify_label(self) -> Self

Align the label to the mid-left of the Button's surface.

This is the default label alignment.

pub fn right_justify_label(self) -> Self

Align the label to the mid-left of the Button's surface.

pub fn label_x(self, x: Relative) -> Self

Specify the label's position relatively to Button along the x axis.

pub fn label_y(self, y: Relative) -> Self

Specify the label's position relatively to Button along the y axis.

pub fn enabled(self, enabled: bool) -> Self

Build the type's self.$($assignee).+ with the given $Type.