Struct conrod::widget::canvas::Canvas

pub struct Canvas<'a> {
    pub common: CommonBuilder,
    pub style: Style,
    pub maybe_title_bar_label: Option<&'a str>,
    pub maybe_splits: Option<FlowOfSplits<'a>>,
}

Canvas is designed to be a "container"-like "parent" widget that simplifies placement of "children" widgets.

Widgets can be placed on a Canvas in a variety of ways using methods from the Positionable trait.

Canvas provides methods for padding the kid widget area which can make using the Place-related Positionable methods a little easier.

A Canvas can also be divided into a sequence of smaller Canvasses using the .flow_* methods. This creates a kind of Canvas tree, where each "split" can be sized using the .length or .length_weight methods.

See the canvas.rs example for a demonstration of the Canvas type.

Fields
common: CommonBuilder

Data necessary and common for all widget builder types.

style: Style

The builder data related to the style of the Canvas.

maybe_title_bar_label: Option<&'a str>

The label for the Canvas' TitleBar if there is one.

maybe_splits: Option<FlowOfSplits<'a>>

A list of child Canvasses as splits of this Canvas flowing in the given direction.

Methods

impl<'a> Canvas<'a>

pub fn new() -> Self

Construct a new Canvas builder.

pub fn title_bar(self, title_bar: &'a str) -> Self

Build the type's self.$($assignee).+ with the given $Type.

pub fn pad_left(self, pad_left: Scalar) -> Self

Build the type's self.$($assignee).+ with the given $Type.

pub fn pad_right(self, pad_right: Scalar) -> Self

Build the type's self.$($assignee).+ with the given $Type.

pub fn pad_bottom(self, pad_bottom: Scalar) -> Self

Build the type's self.$($assignee).+ with the given $Type.

pub fn pad_top(self, pad_top: Scalar) -> Self

Build the type's self.$($assignee).+ with the given $Type.

pub fn with_style(self, with_style: Style) -> Self

Build the type's self.$($assignee).+ with the given $Type.

pub fn length(self, length: Scalar) -> Self

Set the length of the Split as an absolute scalar.

pub fn length_weight(self, weight: Weight) -> Self

Set the length of the Split as a weight.

The default length weight for each widget is 1.0.

pub fn flow_right(self, splits: ListOfSplits<'a>) -> Self

Set the child Canvasses flowing to the right.

pub fn flow_left(self, splits: ListOfSplits<'a>) -> Self

Set the child Canvasses flowing to the left.

pub fn flow_up(self, splits: ListOfSplits<'a>) -> Self

Set the child Canvasses flowing upwards.

pub fn flow_down(self, splits: ListOfSplits<'a>) -> Self

Set the child Canvasses flowing downwards.

pub fn pad(self, pad: Scalar) -> Self

Set the padding for all edges of the area where child widgets will be placed.

pub fn padding(self, pad: Padding) -> Self

Set the padding of the area where child widgets will be placed.

pub fn title_bar_color(self, color: Color) -> Self

Set the color of the Canvas' TitleBar if it is visible.