Struct conrod::widget::canvas::Style

pub struct Style {
    pub color: Option<Color>,
    pub border: Option<Scalar>,
    pub border_color: Option<Color>,
    pub length: Option<Length>,
    pub pad_left: Option<Scalar>,
    pub pad_right: Option<Scalar>,
    pub pad_bottom: Option<Scalar>,
    pub pad_top: Option<Scalar>,
    pub title_bar_color: Option<Option<Color>>,
    pub title_bar_text_color: Option<Color>,
    pub title_bar_font_size: Option<FontSize>,
    pub title_bar_maybe_wrap: Option<Option<Wrap>>,
    pub title_bar_line_spacing: Option<Scalar>,
    pub title_bar_justify: Option<Justify>,
}

Unique styling for the Canvas.

Fields
color: Option<Color>

The color of the Canvas' rectangle surface.

border: Option<Scalar>

The width of the border surrounding the Canvas' rectangle.

border_color: Option<Color>

The color of the Canvas' border.

length: Option<Length>

If this Canvas is a split of some parent Canvas, this is the length of the split.

pad_left: Option<Scalar>

Padding for the left edge of the Canvas' kid area.

pad_right: Option<Scalar>

Padding for the right edge of the Canvas' kid area.

pad_bottom: Option<Scalar>

Padding for the bottom edge of the Canvas' kid area.

pad_top: Option<Scalar>

Padding for the top edge of the Canvas' kid area.

title_bar_color: Option<Option<Color>>

The color of the title bar. Defaults to the color of the Canvas.

title_bar_text_color: Option<Color>

The color of the title bar's text.

title_bar_font_size: Option<FontSize>

The font size for the title bar's text.

title_bar_maybe_wrap: Option<Option<Wrap>>

The way in which the title bar's text should wrap.

title_bar_line_spacing: Option<Scalar>

The distance between lines for multi-line title bar text.

title_bar_justify: Option<Justify>

The label's typographic alignment over the x axis.

Methods

impl Style

fn new() -> Self

Construct the default Style, initialising all fields to None.

impl Style

fn color(&self, theme: &Theme) -> Color

Retrieves the value from the Style.

If the Style's field is None, falls back to default specified within the Theme.

fn border(&self, theme: &Theme) -> Scalar

Retrieves the value from the Style.

If the Style's field is None, falls back to default specified within the Theme.

fn border_color(&self, theme: &Theme) -> Color

Retrieves the value from the Style.

If the Style's field is None, falls back to default specified within the Theme.

fn length(&self, theme: &Theme) -> Length

Retrieves the value from the Style.

If the Style's field is None, falls back to default specified within the Theme.

fn pad_left(&self, theme: &Theme) -> Scalar

Retrieves the value from the Style.

If the Style's field is None, falls back to default specified within the Theme.

fn pad_right(&self, theme: &Theme) -> Scalar

Retrieves the value from the Style.

If the Style's field is None, falls back to default specified within the Theme.

fn pad_bottom(&self, theme: &Theme) -> Scalar

Retrieves the value from the Style.

If the Style's field is None, falls back to default specified within the Theme.

fn pad_top(&self, theme: &Theme) -> Scalar

Retrieves the value from the Style.

If the Style's field is None, falls back to default specified within the Theme.

fn title_bar_color(&self, theme: &Theme) -> Option<Color>

Retrieves the value from the Style.

If the Style's field is None, falls back to default specified within the Theme.

fn title_bar_text_color(&self, theme: &Theme) -> Color

Retrieves the value from the Style.

If the Style's field is None, falls back to default specified within the Theme.

fn title_bar_font_size(&self, theme: &Theme) -> FontSize

Retrieves the value from the Style.

If the Style's field is None, falls back to default specified within the Theme.

fn title_bar_maybe_wrap(&self, theme: &Theme) -> Option<Wrap>

Retrieves the value from the Style.

If the Style's field is None, falls back to default specified within the Theme.

fn title_bar_line_spacing(&self, theme: &Theme) -> Scalar

Retrieves the value from the Style.

If the Style's field is None, falls back to default specified within the Theme.

fn title_bar_justify(&self, theme: &Theme) -> Justify

Retrieves the value from the Style.

If the Style's field is None, falls back to default specified within the Theme.

impl Style

fn padding(&self, theme: &Theme) -> Padding

Get the Padding for the Canvas' kid area.