Struct conrod::widget::drop_down_list::Style

pub struct Style {
    pub color: Option<Color>,
    pub border: Option<Scalar>,
    pub border_color: Option<Color>,
    pub label_color: Option<Color>,
    pub label_font_size: Option<FontSize>,
    pub label_justify: Option<Justify>,
    pub label_x: Option<Relative>,
    pub label_y: Option<Relative>,
    pub maybe_max_visible_height: Option<Option<MaxHeight>>,
    pub scrollbar_position: Option<Option<ScrollbarPosition>>,
    pub scrollbar_width: Option<Option<Scalar>>,
    pub label_font_id: Option<Option<Id>>,
}

Styling for the DropDownList, necessary for constructing its renderable Element.

Fields
color: Option<Color>

Color of the widget.

border: Option<Scalar>

Width of the widget's border.

border_color: Option<Color>

Color of the widget's border.

label_color: Option<Color>

Color of the item labels.

label_font_size: Option<FontSize>

Font size for the item labels.

label_justify: Option<Justify>

The label's typographic alignment over the x axis.

label_x: Option<Relative>

The label's position relative to its Button along the x axis.

label_y: Option<Relative>

The label's position relative to its Button along the y axis.

maybe_max_visible_height: Option<Option<MaxHeight>>

Maximum height of the Open menu before the scrollbar appears.

scrollbar_position: Option<Option<ScrollbarPosition>>

The position of the scrollbar in the case that the list is scrollable.

scrollbar_width: Option<Option<Scalar>>

The width of the scrollbar in the case that the list is scrollable.

label_font_id: Option<Option<Id>>

The ID of the font used to display the labels.

Methods

impl Style

fn new() -> Self

Construct the default Style, initialising all fields to None.

impl Style

fn color(&self, theme: &Theme) -> Color

Retrieves the value from the Style.

If the Style's field is None, falls back to default specified within the Theme.

fn border(&self, theme: &Theme) -> Scalar

Retrieves the value from the Style.

If the Style's field is None, falls back to default specified within the Theme.

fn border_color(&self, theme: &Theme) -> Color

Retrieves the value from the Style.

If the Style's field is None, falls back to default specified within the Theme.

fn label_color(&self, theme: &Theme) -> Color

Retrieves the value from the Style.

If the Style's field is None, falls back to default specified within the Theme.

fn label_font_size(&self, theme: &Theme) -> FontSize

Retrieves the value from the Style.

If the Style's field is None, falls back to default specified within the Theme.

fn label_justify(&self, theme: &Theme) -> Justify

Retrieves the value from the Style.

If the Style's field is None, falls back to default specified within the Theme.

fn label_x(&self, theme: &Theme) -> Relative

Retrieves the value from the Style.

If the Style's field is None, falls back to default specified within the Theme.

fn label_y(&self, theme: &Theme) -> Relative

Retrieves the value from the Style.

If the Style's field is None, falls back to default specified within the Theme.

fn maybe_max_visible_height(&self, theme: &Theme) -> Option<MaxHeight>

Retrieves the value from the Style.

If the Style's field is None, falls back to default specified within the Theme.

fn scrollbar_position(&self, theme: &Theme) -> Option<ScrollbarPosition>

Retrieves the value from the Style.

If the Style's field is None, falls back to default specified within the Theme.

fn scrollbar_width(&self, theme: &Theme) -> Option<Scalar>

Retrieves the value from the Style.

If the Style's field is None, falls back to default specified within the Theme.

fn label_font_id(&self, theme: &Theme) -> Option<Id>

Retrieves the value from the Style.

If the Style's field is None, falls back to default specified within the Theme.

impl Style

fn button_style(&self, is_selected: bool) -> Style

Style for a Button given this Style's current state.