Enum conrod::widget::envelope_editor::Event

pub enum Event<E> where E: EnvelopePoint {
    AddPoint {
        i: usize,
        point: E,
    },
    RemovePoint {
        i: usize,
    },
    MovePoint {
        i: usize,
        x: E::X,
        y: E::Y,
    },
}

The kinds of events that may be yielded by the EnvelopeEditor.

Variants
AddPoint

Insert a new point.

Fields of AddPoint

i: usize

The index at which the point should be inserted.

point: E

The new point.

RemovePoint

Remove a point.

Fields of RemovePoint

i: usize

The index of the point that should be removed.

MovePoint

Move a point.

Fields of MovePoint

i: usize

The index of the point that should be moved.

x: E::X

The point's new x value.

y: E::Y

The point's new y value.

Methods

impl<E> Event<E> where E: EnvelopePoint

fn update(self, envelope: &mut Vec<E>)

Update the given envelope in accordance with the Event.