Module conrod::widget::file_navigator

A widget for navigating through through a file system. Generally inspired by Finder.

Useful for saving widgets that save/load files.

DirectoryView: Lists the contents of a single directory. Reacts to events for selection of one or more files, de-selection, deletion and double-clicking.

FileView: Displays some basic information about the file.

Reexports

pub use self::directory_view::DirectoryView;

Modules

directory_view
Lists the contents of a single directory.

Structs

Directory
Represents the state for a single directory.

FileNavigator
A widget for navigating and interacting with a file system.

State
Unique state stored within the widget graph for each FileNavigator.

Style
Unique styling for the widget.

Enums

Event
The kinds of events that the FileNavigator may produce.

Types
A type for specifying the types of files to be shown by a FileNavigator.