Struct conrod::widget::file_navigator::directory_view::DirectoryView

pub struct DirectoryView<'a> {
    pub style: Style,
    pub directory: &'a Path,
    pub types: Types<'a>,
    // some fields omitted
}

For viewing, selecting, double-clicking, etc the contents of a directory.

Fields
style: Style

Unique styling for the widget.

directory: &'a Path

The path of the directory to display.

types: Types<'a>

Only display files of the given type.

Methods

impl<'a> DirectoryView<'a>

fn new(directory: &'a Path, types: Types<'a>) -> Self

Begin building a DirectoryNavigator widget that displays only files of the given types.

fn unselected_color(self, color: Color) -> Self

The color of the unselected entries within each DirectoryView.

fn text_color(self, color: Color) -> Self

The color of the Text used to display the file names.

fn show_hidden_files(self, show_hidden: bool) -> Self

Whether to show hidden files and directories

fn font_size(self, font_size: FontSize) -> Self

Build the type's self.$($assignee).+ with the given $Type.