Enum conrod::widget::file_navigator::Event

pub enum Event {
    ChangeDirectory(PathBuf),
    ChangeSelection(Vec<PathBuf>),
    Click(Click, Vec<PathBuf>),
    DoubleClick(DoubleClick, Vec<PathBuf>),
    Press(Press, Vec<PathBuf>),
    Release(Release, Vec<PathBuf>),
}

The kinds of events that the FileNavigator may produce.

Variants
ChangeDirectory(PathBuf)

The directory at the top of the stack has changed.

ChangeSelection(Vec<PathBuf>)

The selection of files in the top of the stack has changed.

Click(Click, Vec<PathBuf>)

A Click event occurred over a selection of entries.

DoubleClick(DoubleClick, Vec<PathBuf>)

A file was double clicked.

Press(Press, Vec<PathBuf>)

A Press event occurred over a selection of entries.

Release(Release, Vec<PathBuf>)

A Release event occurred over a selection of entries.