Struct conrod::widget::file_navigator::Style

pub struct Style {
    pub color: Option<Color>,
    pub unselected_color: Option<Option<Color>>,
    pub text_color: Option<Option<Color>>,
    pub font_size: Option<FontSize>,
    pub column_width: Option<Scalar>,
    pub resize_handle_width: Option<Scalar>,
}

Unique styling for the widget.

Fields
color: Option<Color>

Color of the selected entries.

unselected_color: Option<Option<Color>>

The color of the unselected entries.

text_color: Option<Option<Color>>

The color of the directory and file names.

font_size: Option<FontSize>

The font size for the directory and file names.

column_width: Option<Scalar>

The default width of a single directory view.

The first directory will always be initialised to this size.

resize_handle_width: Option<Scalar>

The width of the bar that separates each directory in the stack and allows for re-sizing.

Methods

impl Style

fn new() -> Self

Construct the default Style, initialising all fields to None.

impl Style

fn color(&self, theme: &Theme) -> Color

Retrieves the value from the Style.

If the Style's field is None, falls back to default specified within the Theme.

fn unselected_color(&self, theme: &Theme) -> Option<Color>

Retrieves the value from the Style.

If the Style's field is None, falls back to default specified within the Theme.

fn text_color(&self, theme: &Theme) -> Option<Color>

Retrieves the value from the Style.

If the Style's field is None, falls back to default specified within the Theme.

fn font_size(&self, theme: &Theme) -> FontSize

Retrieves the value from the Style.

If the Style's field is None, falls back to default specified within the Theme.

fn column_width(&self, theme: &Theme) -> Scalar

Retrieves the value from the Style.

If the Style's field is None, falls back to default specified within the Theme.

fn resize_handle_width(&self, theme: &Theme) -> Scalar

Retrieves the value from the Style.

If the Style's field is None, falls back to default specified within the Theme.