Function conrod::widget::graph::edge_node_widget_ids

pub fn edge_node_widget_ids<NI>(edge: &Edge<NI>, ui: &Ui) -> (Id, Id) where
    NI: NodeId, 

Returns the widget::Ids for the start and end nodes.

Edges can only exist for the lifetime of a SessionEdges, thus we assume that there will always be a Graph for the edge's graph_id and that there will always be a widget::Id for the start and end nodes.

Panic!s if the given Ui is not the same one used to create the edge's parent Graph widget.