Struct conrod::widget::graph::node::Node

pub struct Node<W> {
    pub common: CommonBuilder,
    pub style: Style,
    pub widget: W,
    pub inputs: usize,
    pub outputs: usize,
}

A widget that acts as a convenience container for some Node's unique widgets.

Fields
common: CommonBuilder

Data necessary and common for all widget builder types.

style: Style

Unique styling for the Node.

widget: W

The widget wrapped by this node container.

inputs: usize

The number of input sockets on the node.

outputs: usize

The number of output sockets on the node.

Methods

impl<W> Node<W>

pub fn new(widget: W) -> Self

Begin building a new Node widget.

pub fn inputs(self, inputs: usize) -> Self

Specify the number of input sockets for the node.

pub fn outputs(self, outputs: usize) -> Self

Specify the number of output sockets for the node.

pub fn color(self, color: Color) -> Self

Specify the color for the node's inner rectangle.

pub fn border_thickness(self, thickness: Scalar) -> Self

The thickness of the border around the inner widget.

This must always be a positive value in order for sockets to remain visible.

pub fn border_color(self, color: Color) -> Self

Specify the color for the node's border.

pub fn border_radius(self, radius: Scalar) -> Self

Specify the radius for the node's border.

pub fn socket_color(self, color: Color) -> Self

Specify the color for the node's sockets.

pub fn input_socket_layout(self, layout: SocketLayout) -> Self

Specify the layout of the input sockets.

pub fn output_socket_layout(self, layout: SocketLayout) -> Self

Specify the layout of the input sockets.