Struct conrod::widget::graph::Edge

pub struct Edge<'a, NI: NodeId> { /* fields omitted */ }

A context for an edge yielded during the edge instantiation stage.

Tyis type can

Methods

impl<'a, NI> Edge<'a, NI> where
NI: NodeId,

pub fn start(&self) -> NodeSocket<NI>

The start (or "input") for the edge.

This is described via the node's Id and the position of its output socket.

pub fn end(&self) -> NodeSocket<NI>

The end (or "output") for the edge.

This is described via the node's Id and the position of its input socket.

pub fn sockets(&self) -> (NodeSocket<NI>, NodeSocket<NI>)

The start and end sockets.

pub fn widget<W>(self, widget: W) -> EdgeWidget<'a, NI, W>

Specify the widget to use