Struct conrod::widget::graph::Layout

pub struct Layout<NI> where
    NI: Eq + Hash,  { /* fields omitted */ }

Stores the layout of all nodes within the graph.

All positions are relative to the centre of the Graph widget.

Nodes can be moved by