Struct conrod::widget::graph::NodeContext

pub struct NodeContext<'a, NI: 'a + NodeId> { /* fields omitted */ }

A context for a node yielded during the node instantiation stage.

This type can be used to:

Get the position of the node via point().

Get the ID for this node via node_id().

Convert into a NodeWidget ready for instantiation within the Ui via widget(a_widget).

Methods

impl<'a, NI> NodeContext<'a, NI> where
NI: NodeId,

pub fn node_id(&self) -> NI

The unique identifier associated with this node.

pub fn point(&self) -> Point

The location of the node.

pub fn widget<W>(self, widget: W) -> NodeWidget<'a, NI, W>

Specify the widget to use