Struct conrod::widget::graph::NodeWidget

pub struct NodeWidget<'a, NI: 'a + NodeId, W> { /* fields omitted */ }

Returned when a NodeContext is assigned a widget.

This intermediary type allows for accessing the widget::Id before the widget itself is instantiated.

Methods

impl<'a, NI, W> NodeWidget<'a, NI, W> where
NI: NodeId,
W: 'static + Widget,

pub fn widget_id(&self, ui: &mut UiCell) -> Id

Retrieve the widget::Id that will be used to instantiate this node's widget.

pub fn map<M>(self, map: M) -> Self where
M: FnOnce(W) -> W,

Map over the inner widget.

pub fn set(self, ui: &mut UiCell) -> W::Event

Set the given widget for the node at node_id().