Struct conrod::widget::graph::Session

pub struct Session<NI: NodeId> { /* fields omitted */ }

A context for moving through the modes of graph widget instantiation in a type-safe manner.

The Session is shared between 3 stages:

SessionEvents: Emit all graph events that have occurred since the last instantiation.

SessionNodes: Instantiate all node widgets in the graph.

SessionEdges: Instantiate all edge widgets in the graph.

NOTE: This should allow for different instantiation orders, e.g: nodes then edges, all connected components in topo order, edges then nodes, etc.