Struct conrod::widget::grid::Lines

pub struct Lines<T> {
    pub step: T,
    pub offset: Option<T>,
    pub thickness: Option<Scalar>,
    pub color: Option<Color>,
}

A series of lines distributed across an axis.

Fields
step: T

The distance that separates each line.

offset: Option<T>

An optional offset for the lines along they're axis.

thickness: Option<Scalar>

The thickness of each of the lines drawn.

If None, the thickness specified within the Style is used.

color: Option<Color>

The color of each of the lines drawn.

If None, the color specified within the Style is used.

Methods

impl<T> Lines<T>

fn step(step: T) -> Self

Begin building a new set of lines for the grid step distance apart.

Lines with a step that equates to 0.0 or less will not be drawn.

fn offset(self, offset: T) -> Self

Specify an offset for the grid.

Offsets that are greater than the step size will be wrapped around the step size.

fn thickness(self, thickness: Scalar) -> Self

Specify a unique thickness for these lines.

fn color(self, color: Color) -> Self

Use the specified color to uniquely color the this set of lines.

fn x<Y>(self) -> Axis<T, Y>

Move the lines over the X axis.

fn y<X>(self) -> Axis<X, T>

Move the lines over the Y axis.