Module conrod::widget::list

A helper widget, useful for instantiating a sequence of widgets in a vertical list.

Structs

Dynamic
A type that implements ItemSize for Lists whose Items' length are unknown until setting the widget for each item.

Fixed
A type that implements ItemSize for Lists whose Items are a fixed size and known prior to setting the widgets for each item.

Item
The data necessary for instantiating a single item within a List.

Items
An Iterator yielding each Item in the list.

List
A helper widget, useful for instantiating a sequence of widgets in a vertical list.

Scrollbar
A wrapper around a List's Scrollbar and its widget::Id.

State
Represents the state of the List widget.

Style
Unique styling for the List.

Enums

Down
Items flow from top to bottom.

ItemInstantiation
The way in which a List should instantiate its Items.

Left
Items flow from right to left.

Right
Items flow from left to right.

ScrollbarPosition
If the List is scrollable, this describes how th Scrollbar should be positioned.

Up
Items flow from bottom to top.

Traits

Direction
The direction in which the list is laid out.

ItemSize
The way in which the List's items are sized. E.g. Fired or Dynamic.