Struct conrod::widget::list::Item

pub struct Item {
    pub i: usize,
    pub widget_id: Id,
    pub last_id: Option<Id>,
    pub w: Scalar,
    pub h: Scalar,
    // some fields omitted
}

The data necessary for instantiating a single item within a List.

Fields
i: usize

The index of the item within the list.

widget_id: Id

The id generated for the widget.

last_id: Option<Id>

The id used for the previous item's widget.

w: Scalar

The width of the item.

h: Scalar

The height of the item.

Methods

impl Item

pub fn set<W>(self, widget: W, ui: &mut UiCell) -> W::Event where
W: Widget,

Sets the given widget as the widget to use for the item.

Sets the:

position of the widget.

dimensions of the widget.

parent of the widget.

and finally sets the widget within the Ui.