Struct conrod::widget::list::List

pub struct List<D, S> {
    pub common: CommonBuilder,
    pub style: Style,
    pub item_instantiation: ItemInstantiation,
    // some fields omitted
}

A helper widget, useful for instantiating a sequence of widgets in a vertical list.

The List widget simplifies this process by:

Generating widget::Ids.

Simplifying the positioning and sizing of items.

Optimised widget instantiation by only instantiating visible items. This is very useful for lists containing many items, i.e. a FileNavigator over a directory with thousands of files.

Fields
common: CommonBuilder

Common widget building params for the List.

style: Style

Unique styling for the List.

item_instantiation: ItemInstantiation

Whether all or only visible items should be instantiated.

Methods

impl<D> List<D, Dynamic> where
D: Direction,

pub fn new(num_items: usize) -> Self

Begin building a new List.

impl List<Left, Dynamic>

pub fn flow_left(num_items: usize) -> Self

Begin building a new List flowing from right to left.

impl List<Right, Dynamic>

pub fn flow_right(num_items: usize) -> Self

Begin building a new List flowing from left to right.

impl List<Up, Dynamic>

pub fn flow_up(num_items: usize) -> Self

Begin building a new List flowing from bottom to top.

impl List<Down, Dynamic>

pub fn flow_down(num_items: usize) -> Self

Begin building a new List flowing from top to bottom.

impl<D, S> List<D, S> where
D: Direction,
S: ItemSize,

pub fn from_item_size(num_items: usize, item_size: S) -> Self

Begin building a new List given some direction and item size.

pub fn item_size(self, length: Scalar) -> List<D, Fixed>

Specify a fixed item size, where size is a Scalar in the direction that the List is flowing. When a List is constructed with this method, all items will have a fixed, equal length.

impl<D> List<D, Fixed> where
D: Direction,

pub fn instantiate_all_items(self) -> Self

Indicates that an Item should be instatiated for every element in the list, regardless of whether or not the Item would be visible.

This is the default (and only) behaviour for Lists with dynamic item sizes. This is because a List cannot know the total length of its combined items in advanced when each item is dynamically sized and their size is not given until they are set.

Note: This may cause significantly heavier CPU load for lists containing many items (100+). We only recommend using this when absolutely necessary as large lists may cause unnecessary bloating within the widget graph, and in turn result in greater traversal times.

pub fn instantiate_only_visible_items(self) -> Self

Indicates that only Items that are visible should be instantiated. This ensures that we avoid bloating the widget graph with unnecessary nodes and in turn keep traversal times to a minimum.

This is the default behaviour for Lists with fixed item sizes.

impl<D, S> List<D, S> where
D: Direction,
S: ItemSize,

pub fn scrollbar_next_to(self) -> Self

Specifies that the List should be scrollable and should provide a Scrollbar to the right of the items.

pub fn scrollbar_on_top(self) -> Self

Specifies that the List should be scrollable and should provide a Scrollbar that hovers above the right edge of the items and automatically hides when the user is not scrolling.

pub fn scrollbar_thickness(self, w: Scalar) -> Self

The width of the Scrollbar.

pub fn scrollbar_color(self, color: Color) -> Self

The color of the Scrollbar.