Module conrod::widget::list_select

A wrapper around the List widget providing the ability to select one or more items.

Structs

Events
An iterator-like type for yielding ListSelect Events.

ListSelect
A wrapper around the List widget that handles single and multiple selection logic.

Multiple
A selection Mode for the ListSelect that allows selecting more than one item at a time.

Single
A single item selection Mode for the ListSelect.

State
Represents the state of the ListSelect.

Enums

Event
The kind of events that the ListSelect may react to. Provides tuple(s) of index in list and string representation of selection

Selection
Represents some change in item selection for a ListSelect in Multiple mode.

Traits

Mode
Allows the ListSelect to be generic over Single and Multiple selection modes.

Type Definitions

PendingEvents
Buffer use for storing events that have been produced but are yet to be yielded.