Enum conrod::widget::list_select::Event

pub enum Event<S> {
    Item(Item),
    Selection(S),
    Press(Press),
    Release(Release),
    Click(Click),
    DoubleClick(DoubleClick),
}

The kind of events that the ListSelect may react to. Provides tuple(s) of index in list and string representation of selection

Variants
Item(Item)

The next Item is ready for instantiation.

Selection(S)

A change in selection has occurred.

Press(Press)

A button press occurred while the widget was capturing the mouse.

Release(Release)

A button release occurred while the widget was capturing the mouse.

Click(Click)

A click occurred while the widget was capturing the mouse.

DoubleClick(DoubleClick)

A double click occurred while the widget was capturing the mouse.