Enum conrod::widget::list_select::Selection

pub enum Selection<H: BuildHasher = RandomState> {
    Add(HashSet<usize, H>),
    Remove(HashSet<usize, H>),
}

Represents some change in item selection for a ListSelect in Multiple mode.

Variants
Add(HashSet<usize, H>)

Items which have been added to the selection.

Remove(HashSet<usize, H>)

Items which have been removed from the selection.

Methods

impl<H: BuildHasher> Selection<H>

pub fn update_bool_slice(&self, slice: &mut [bool])

Update the given slice of bools with this Selection.

Each index in the Selection represents and index into the slice.

pub fn update_index_set<T>(&self, set: &mut HashSet<usize, T>) where
T: BuildHasher,

Update the given set of selected indices with this Selection.