Module conrod::widget::matrix

A helper widget for laying out child widgets in the form of a grid.

Structs

Element
Data necessary for instantiating a widget for a single Matrix element.

Elements
The event type yielded by the Matrix.

Matrix
Draw a matrix of any rectangular widget type, where the matrix will provide a function with the widget number, it's rows and cols position, the width and height for the widget and the location at which the widget should be drawn.

State
The state of the Matrix, to be cached within the Ui's widget Graph.

Style
Unique styling for the Matrix.

Type Definitions

ColNum
A column index.

Height
The height of an element.

PosX
The position of an element along the x axis.

PosY
The position of an element along the y axis.

RowNum
A row index.

WidgetNum
The number of the widget.

Width
The width of an element.