Struct conrod::widget::matrix::Matrix

pub struct Matrix { /* fields omitted */ }

Draw a matrix of any rectangular widget type, where the matrix will provide a function with the widget number, it's rows and cols position, the width and height for the widget and the location at which the widget should be drawn.

Methods

impl Matrix

pub fn new(cols: usize, rows: usize) -> Self

Create a widget matrix context.

pub fn cell_padding(self, w: Scalar, h: Scalar) -> Self

A builder method for adding padding to the cell.