Struct conrod::widget::plot_path::PlotPath

pub struct PlotPath<X, Y, F> { /* fields omitted */ }

A widget that plots a series of lines using the given function x -> y.

The function is sampled once per pixel and the result is mapped to the widget's height.

The resulting "path" is drawn using conrod's PointPath primitive widget.

Methods

impl<X, Y, F> PlotPath<X, Y, F>

pub fn new(min_x: X, max_x: X, min_y: Y, max_y: Y, f: F) -> Self

Begin building a new PlotPath widget instance.

pub fn thickness(self, thickness: Scalar) -> Self

The thickness of the point path used to draw the plot.