Struct conrod::widget::primitive::image::Image

pub struct Image {
    pub common: CommonBuilder,
    pub image_id: Id,
    pub src_rect: Option<Rect>,
    pub style: Style,
}

A primitive and basic widget for drawing an Image.

Fields
common: CommonBuilder

Data necessary and common for all widget builder types.

image_id: Id

The unique identifier for the image that will be drawn.

src_rect: Option<Rect>

The rectangle area of the original source image that should be used.

style: Style

Unique styling.

Methods

impl Image

pub fn new(image_id: Id) -> Self

Construct a new Image.

Note that the Image widget does not require borrowing or owning any image data directly. Instead, image data is stored within a conrod::image::Map where image::Ids are mapped to their associated data.

This is done for a few reasons:

To avoid requiring that the widget graph owns an instance of each image

To avoid requiring that the user passes the image data to the Image every update unnecessarily

To make it easier for users to borrow and mutate their images without needing to index into the Ui's widget graph (which also requires casting types).

During rendering, conrod will take the image::Map, retrieve the data associated with each image and yield it via the render::Primitive::Image variant.

Note: this implies that the type must be the same for all Image widgets instantiated via the same Ui. In the case that you require multiple different types of images, we recommend that you either:

use an enum with a variant for each type

use a trait object, where the trait is implemented for each of your image types or

use an index type which may be mapped to your various image types.

pub fn source_rectangle(self, rect: Rect) -> Self

The rectangular area of the image that we wish to display.

If this method is not called, the entire image will be used.

pub fn color(self, color: Option<Color>) -> Self

Build the type's self.$($assignee).+ with the given $Type.