Module conrod::widget::primitive::line

A simple, non-interactive widget for drawing a single straight Line.

Structs

Line
A simple, non-interactive widget for drawing a single straight Line.

State
Unique state for the Line widget.

Style
Unique styling for a Line widget.

Enums

Cap
Whether the end of the Line should be flat or rounded.

Pattern
The pattern used to draw the line.

Functions

is_over
Describes whether or not the given point touches the line described by a -> b with the given thickness.

is_over_widget
The function to use for picking whether a given point is over the line.

rect_corners
Given two points and half the line thickness, return the four corners of the rectangle describing the line.

triangles
Given two points and half the line thickness, return the two triangles that describe the line.