Struct conrod::widget::primitive::shape::triangles::Triangles

pub struct Triangles<S, I> {
    pub common: CommonBuilder,
    pub style: S,
    pub triangles: I,
    pub maybe_shift_to_centre_from: Option<Point>,
}

A widget that allows for drawing a list of triangles.

Fields
common: CommonBuilder

Data necessary and common for all widget builder types.

style: S

Unique styling for the Triangles.

triangles: I

All the point in the triangle list.

maybe_shift_to_centre_from: Option<Point>

Whether or not the triangles should be automatically centred to the widget position.

Methods

impl<I> Triangles<SingleColor, I> where
I: IntoIterator<Item = Triangle<<SingleColor as Style>::Vertex>>,

pub fn single_color<C>(
color: C,
points: I
) -> TrianglesUnpositioned<SingleColor, I> where
C: Into<Rgba>,

A list of triangles described by the given points.

All triangles are colored with the given Color.

impl<I> Triangles<MultiColor, I> where
I: IntoIterator<Item = Triangle<<MultiColor as Style>::Vertex>>,

pub fn multi_color(points: I) -> TrianglesUnpositioned<MultiColor, I>

A list of triangles described by the given points.

Every vertex specifies its own unique color.