Struct conrod::widget::CommonBuilder

pub struct CommonBuilder {
    pub style: CommonStyle,
    pub maybe_parent_id: MaybeParent,
    pub is_floating: bool,
    pub crop_kids: bool,
    pub maybe_x_scroll: Option<Scroll>,
    pub maybe_y_scroll: Option<Scroll>,
    pub place_on_kid_area: bool,
    pub maybe_graphics_for: Option<Id>,
}

A struct containing builder data common to all Widget types.

This type also allows us to do a blanket impl of Positionable and Sizeable for T: Widget.

When Rust gets some sort of field inheritance feature, this will most likely be refactored to take advantage of that.

Fields
style: CommonStyle

Styling and positioning data that is common between all widget types.

maybe_parent_id: MaybeParent

The parent widget of the Widget.

is_floating: bool

Whether or not the Widget is a "floating" Widget.

crop_kids: bool

Whether or not the children of this Widget should be cropped to its kid_area.

By default, the kid_area is the size of the entire widget, though it may be specified otherwise via the Widget::kid_area method.

maybe_x_scroll: Option<Scroll>

Arguments to the scrolling of the widget's x axis.

maybe_y_scroll: Option<Scroll>

Arguments to the scrolling of the widget's y axis.

place_on_kid_area: bool

Whether or not the Widget should be placed on the kid_area.

If true, the Widget will be placed on the kid_area of the parent Widget if the Widget is given a Place variant for its Position.

If false, the Widget will be placed on the parent Widget's total area.

maybe_graphics_for: Option<Id>

Describes whether or not the Widget is instantiated as a graphical element for some other Widget.

When adding an edge a -> b where b is considered to be a graphical element of a, several things are implied about b:

If b is picked within either Graph::pick_widget or Graph::pick_top_scrollable_widget, it will instead return the index for a.

When determining the Graph::scroll_offset for b, a's scrolling (if it is scrollable, that is) will be skipped.

b will always be placed upon a's total area, rather than its kid_area which is the default.

Any Graphic child of b will be considered as a Graphic child of a.