Struct conrod::widget::Id

pub struct Id(pub usize);

Unique, public widget identifier. Each widget must use a unique WidgetId so that it's state can be cached within the Ui type. The reason we use a usize is because widgets are cached within a Graph whose max number of Nodes is indexed by usize.