Struct conrod::widget::PreUpdateCache

pub struct PreUpdateCache {
    pub type_id: TypeId,
    pub id: Id,
    pub maybe_parent_id: Option<Id>,
    pub maybe_x_positioned_relatively_id: Option<Id>,
    pub maybe_y_positioned_relatively_id: Option<Id>,
    pub rect: Rect,
    pub depth: Depth,
    pub kid_area: KidArea,
    pub maybe_floating: Option<Floating>,
    pub crop_kids: bool,
    pub maybe_x_scroll_state: Option<StateX>,
    pub maybe_y_scroll_state: Option<StateY>,
    pub maybe_graphics_for: Option<Id>,
}

Fields
type_id: TypeId

The Widget's unique type identifier.

id: Id

The Widget's unique Id.

maybe_parent_id: Option<Id>

The Widget's parent's unique index (if it has a parent).

maybe_x_positioned_relatively_id: Option<Id>

If this Widget is relatively positioned to another Widget, this will be the index of the Widget to which this Widget is relatively positioned along the x axis.

maybe_y_positioned_relatively_id: Option<Id>

If this Widget is relatively positioned to another Widget, this will be the index of the Widget to which this Widget is relatively positioned along the y axis.

rect: Rect

The Rect describing the Widget's position and dimensions.

depth: Depth

The z-axis depth - affects the render order of sibling widgets.

kid_area: KidArea

The area upon which the Widget's children widgets will be placed.

maybe_floating: Option<Floating>

Floating data for the Widget if there is some.

crop_kids: bool

Whether or not the children of the Widget should be cropped to its kid_area.

maybe_x_scroll_state: Option<StateX>

Scrolling data for the Widget's x axis if there is some.

maybe_y_scroll_state: Option<StateY>

Scrolling data for the Widget's y axis if there is some.

maybe_graphics_for: Option<Id>

Whether or not the Widget has been instantiated as a graphical element for some other widget.