Module conrod::widget::tabs

A wrapper around a list of Canvasses that displays them as a list of selectable tabs.

Structs

State
The state to be cached within the Canvas.

Style
Unique styling for the Tabs widget.

Tab
A single Tab in the list owned by the Tabs State.

TabRects
An iterator yielding the Rect for each Tab in the given list.

Tabs
A wrapper around a list of Canvasses that displays them as a list of selectable tabs.

Enums

Layout
The direction in which the tabs will be laid out.