Struct conrod::widget::tabs::Tabs

pub struct Tabs<'a> {
    // some fields omitted
}

A wrapper around a list of Canvasses that displays them as a list of selectable tabs.

Methods

impl<'a> Tabs<'a>

fn new(tabs: &'a [(Id, &'a str)]) -> Tabs<'a>

Construct some new Canvas Tabs.

fn starting_canvas(self, canvas_id: Id) -> Self

Set the initially selected tab with a Canvas via its widget::Id.

fn pad(self, pad: Scalar) -> Tabs<'a>

Set the padding for all edges.

fn layout_horizontally(self) -> Self

Layout the tabs horizontally.

fn layout_vertically(self) -> Self

Layout the tabs vertically.

fn canvas_style(self, style: Style) -> Self

Build the Tabs widget with the given styling for its Canvasses.

fn pad_left(self, pad: Scalar) -> Self

If the Tabs has some widget::canvas::Style, assign the left padding.

fn pad_right(self, pad: Scalar) -> Self

If the Tabs has some widget::canvas::Style, assign the left padding.

fn pad_bottom(self, pad: Scalar) -> Self

If the Tabs has some widget::canvas::Style, assign the left padding.

fn pad_top(self, pad: Scalar) -> Self

If the Tabs has some widget::canvas::Style, assign the left padding.

fn bar_thickness(self, thickness: Scalar) -> Self

The width of a vertical Tabs selection bar, or the height of a horizontal one.

fn starting_tab_idx(self, starting_tab_idx: usize) -> Self

Build the type's self.$($assignee).+ with the given $Type.

fn label_color(self, label_color: Color) -> Self

Build the type's self.$($assignee).+ with the given $Type.

fn label_font_size(self, label_font_size: FontSize) -> Self

Build the type's self.$($assignee).+ with the given $Type.