Struct conrod::widget::toggle::Toggle

pub struct Toggle<'a> {
    pub enabled: bool,
    // some fields omitted
}

A pressable widget for toggling the state of a bool.

Like the Button widget, it's reaction is triggered upon release and will return the new bool state.

Note that the Toggle will not mutate the bool for you, you should do this yourself within the react function.

Fields
enabled: bool

If true, will allow user inputs. If false, will disallow user inputs.

Methods

impl<'a> Toggle<'a>

fn new(value: bool) -> Toggle<'a>

Construct a new Toggle widget.

fn enabled(self, enabled: bool) -> Self

Build the type's self.$($assignee).+ with the given $Type.