Function conrod::backend::piston::gfx::draw_primitive

pub fn draw_primitive<'a, 'b, Img, F>(context: Context,
                                      graphics: &'a mut G2d<'b>,
                                      primitive: Primitive,
                                      glyph_cache: &'a mut GlyphCache,
                                      image_map: &'a Map<Img>,
                                      glyph_rectangles: &mut Vec<([f64; 4], [f64; 4])>,
                                      texture_from_image: F) where F: FnMut(&Img) -> &G2dTexture<'static>

Draw a single Primitive to the screen.

This is useful if the user requires rendering primitives individually, perhaps to perform their own rendering in between, etc.