Enum conrod::event::Widget

pub enum Widget {
    Text(Text),
    Move(Move),
    Press(Press),
    Release(Release),
    Click(Click),
    DoubleClick(DoubleClick),
    Drag(Drag),
    Scroll(Scroll),
    WindowResized(Dimensions),
    CapturesMouse,
    UncapturesMouse,
    CapturesKeyboard,
    UncapturesKeyboard,
}

Events that apply to a specific widget.

Rather than delivering entire event::Events to the widget (with a lot of redundant information), this event::Widget is used as a refined, widget-specific event.

All Widget event co-ordinates will be relative to the centre of the Widget to which they are delivered.

Variants
Text(Text)

Entered text.

Move(Move)

Represents all forms of motion input.

Press(Press)

Some button was pressed.

Release(Release)

Some button was released.

Click(Click)

Represents a pointing device being pressed and subsequently released while over the same location.

DoubleClick(DoubleClick)

Two Click events with the same button and xy occurring within a duration that is less that the theme.double_click_threshold.

Drag(Drag)

Represents a pointing device button being pressed and a subsequent movement of the mouse.

Scroll(Scroll)

Represents the amount of scroll that has been applied to this widget.

WindowResized(Dimensions)

The window's dimensions were resized.

CapturesMouse

The widget has captured the mouse.

UncapturesMouse

The widget has released the mouse from capturing.

CapturesKeyboard

The widget has captured the keyboard.

UncapturesKeyboard

Indicates that the given widget has released the keyboard from capturing.