Struct conrod::graph::Container

pub struct Container {
    pub maybe_state: Option<Box<Any + Send>>,
    pub type_id: TypeId,
    pub rect: Rect,
    pub depth: Depth,
    pub kid_area: KidArea,
    pub maybe_floating: Option<Floating>,
    pub crop_kids: bool,
    pub maybe_x_scroll_state: Option<StateX>,
    pub maybe_y_scroll_state: Option<StateY>,
    pub instantiation_order_idx: usize,
    pub is_over: IsOverFn,
}

A container for caching a Widget's state inside a Graph Node.

Fields
maybe_state: Option<Box<Any + Send>>

Dynamically stored widget state.

type_id: TypeId

The unique TypeId associated with the Widget::State.

This is equal to std::any::TypeId::of::<Widget::State>().

rect: Rect

The rectangle describing the Widget's area.

depth: Depth

The depth at which the widget will be rendered comparatively to its siblings.

kid_area: KidArea

The area in which child widgets are placed.

maybe_floating: Option<Floating>

Whether or not the widget is a "Floating" widget.

See the Widget::float docs for an explanation of what this means.

crop_kids: bool

Whether or not children widgets should be cropped to the kid_area.

maybe_x_scroll_state: Option<StateX>

Scroll related state (is only Some if this axis is scrollable).

maybe_y_scroll_state: Option<StateY>

Scroll related state (is only Some if this axis is scrollable).

instantiation_order_idx: usize

Represents the Widget's position within the overall instantiation ordering of the widgets.

i.e. if foo's instantiation_order_idx is lower than bar's, it means that foo was instantiated before bar.

is_over: IsOverFn

A function specified by the widget to use when determining whether or not a point is over it.

NOTE: See Wiget::is_over for more details and a note on possible future plans.

Methods

impl Container

pub fn state_and_style<State, Style>(
&self
) -> Option<&UniqueWidgetState<State, Style>> where
State: Any + 'static,
Style: Any + 'static,

Borrow the Container's unique widget State and Style if there is any.

pub fn unique_widget_state<W>(
&self
) -> Option<&UniqueWidgetState<W::State, W::Style>> where
W: Widget,
W::State: Any + 'static,
W::Style: Any + 'static,

Same as Container::state_and_style but accessed using a Widget type parameter instead of the unique State and Style types.