Struct conrod::input::global::Global

pub struct Global {
    pub start: State,
    pub current: State,
    pub last_click: Option<(Instant, Click)>,
    // some fields omitted
}

Global input event handler that also implements input::Provider. The Ui passes all events to it's Global instance, which aggregates and interprets the events to provide so-called 'high-level' events to widgets. This input gets reset after every update by the Ui.

Fields
start: State

The input::State as it was at the end of the last update cycle.

current: State

The most recent input::State, with updates from handling all the events this update cycle

last_click: Option<(Instant, Click)>

Tracks the last click that occurred and the time at which it occurred in order to create double-click events.

Methods

impl Global

fn new() -> Global

Returns a fresh new Global

fn events(&self) -> Events

Returns an iterator yielding all events that have occurred since the last time Ui::set_widgets was called.

fn push_event(&mut self, event: Event)

Add the new event to the stack.

fn clear_events_and_update_start_state(&mut self)

Called at the end of every update cycle in order to prepare the Global to handle events for the next one.