Struct conrod::Canvas

pub struct Canvas {
    pub kind: Kind,
    pub xy: Point,
    pub element: Element,
    pub widget_area_xy: Point,
    pub widget_area_dim: Dimensions,
    pub padding: Padding,
    pub has_updated: bool,
}

The kind of Canvas.

Fields

kind

The Kind of Canvas (i.e. Split, Floating, etc).

xy

The position of the Canvas within the window.

element

The Element used for drawing the Canvas.

widget_area_xy

The position of the Widget area.

widget_area_dim

The dimensions of the Widget area.

padding

Padding for the Canvas describes the distance between each edge and its Widget's.

has_updated

Has the Canvas been set since the last time the Ui was drawn?

Methods

impl Canvas

fn empty() -> Canvas

An empty Canvas