Struct conrod::Polygon

pub struct Polygon<I> {
    pub points: I,
    pub common: CommonBuilder,
    pub style: Style,
    pub maybe_shift_to_centre_from: Option<Point>,
}

A basic, non-interactive, arbitarry Polygon widget.

The Polygon is described by specifying its corners in order.

Polygon will automatically close all shapes, so the given list of points does not need to start and end with the same position.

Fields
points: I

The points describing the corners of the Polygon.

common: CommonBuilder

Data necessary and common for all widget builder types.

style: Style

Unique styling for the Polygon.

maybe_shift_to_centre_from: Option<Point>

Whether or not the points should be automatically centred to the widget position.

Methods

impl<I> Polygon<I>

fn styled(points: I, style: Style) -> Self

Build a polygon with the given points and style.

fn fill(points: I) -> Self

Build a Polygon with the default Fill style.

fn fill_with(points: I, color: Color) -> Self

Build a Polygon Filled with the given Color.

fn outline(points: I) -> Self

Build a Polygon with the default Outline style.

fn outline_styled(points: I, style: LineStyle) -> Self

Build a Polygon Outlineed with the given line style.

fn abs_styled(points: I, style: Style) -> Self where I: IntoIterator<Item=Point> + Clone

Build a new filled Polygon whose bounding box is fit to the absolute co-ordinates of the points.

This requires that the points iterator is Clone so that we may iterate through and determine the bounding box of the points.

If you would rather centre the points to the middle of the bounding box, use the Polygon::centred methods instead.

fn abs_fill(points: I) -> Self where I: IntoIterator<Item=Point> + Clone

The same as Polygon::abs_styled but builds the Polygon with the default Fill style.

fn abs_fill_with(points: I, color: Color) -> Self where I: IntoIterator<Item=Point> + Clone

The same as Polygon::abs_styled but builds the Polygon Filled with the given Color.

fn abs_outline(points: I) -> Self where I: IntoIterator<Item=Point> + Clone

The same as Polygon::abs_styled but builds the Polygon with the default Outline style.

fn abs_outline_styled(points: I, style: LineStyle) -> Self where I: IntoIterator<Item=Point> + Clone

The same as Polygon::abs_styled but builds the Polygon with the given Outline styling.

fn centred_styled(points: I, style: Style) -> Self where I: IntoIterator<Item=Point> + Clone

Build a new Polygon and shift the location of the points so that the centre of their bounding rectangle lies at the position determined for the Polygon widget.

This is useful if your points simply describe a shape and you want to position them using conrod's auto-layout and/or Positionable methods.

If you would rather centre the bounding box to the points, use the Polygon::abs constructor method instead.

fn centred_fill(points: I) -> Self where I: IntoIterator<Item=Point> + Clone

The same as Polygon::centred_styled but constructs the Polygon with the default Fill style.

fn centred_fill_with(points: I, color: Color) -> Self where I: IntoIterator<Item=Point> + Clone

The same as Polygon::centred_styled but constructs the Polygon Filled with the given color.

fn centred_outline(points: I) -> Self where I: IntoIterator<Item=Point> + Clone

The same as Polygon::centred_styled but constructs the Polygon with the default Outline style.

fn centred_outline_styled(points: I, style: LineStyle) -> Self where I: IntoIterator<Item=Point> + Clone

The same as Polygon::centred_styled but constructs the Polygon Outlined with the given styling.