Struct conrod::Rectangle

pub struct Rectangle {
    pub common: CommonBuilder,
    pub style: Style,
}

A basic, non-interactive rectangle shape widget.

Fields

common

Data necessary and common for all widget builder types.

style

Unique styling for the Rectangle.

Methods

impl Rectangle

fn styled(dim: Dimensions, style: Style) -> Self

Build a rectangle with the dimensions and style.

fn fill(dim: Dimensions) -> Self

Build a new filled rectangle.

fn fill_with(dim: Dimensions, color: Color) -> Self

Build a new filled rectangle widget filled with the given color.

fn outline(dim: Dimensions) -> Self

Build a new outlined rectangle widget.

fn outline_styled(dim: Dimensions, line_style: LineStyle) -> Self

Build an outlined rectangle rather than a filled one.