Struct conrod::Ui

pub struct Ui<C> {
    pub theme: Theme,
    pub glyph_cache: GlyphCache<C>,
    pub win_w: f64,
    pub win_h: f64,
    // some fields omitted
}

Ui is the most important type within Conrod and is necessary for rendering and maintaining widget state.

Ui Handles the following:

Contains the state of all widgets which can be indexed via their UiId.

Stores rendering state for each widget until the end of each render cycle.

Contains the theme used for default styling of the widgets.

Maintains the latest user input state (for mouse and keyboard).

Maintains the latest window dimensions.

Fields

theme

The theme used to set default styling for widgets.

glyph_cache

Cache for character textures, used for label width calculation and glyph rendering.

win_w

Window width.

win_h

Window height.

Methods

impl<C> Ui<C>

fn new(character_cache: C, theme: Theme) -> Ui<C>

Constructor for a UiContext.

fn widget_size(&self, id: WidgetId) -> Dimensions

Return the dimensions of a Canvas.

fn canvas_size(&self, id: CanvasId) -> Dimensions

Return the dimensions of a Canvas.

fn handle_event<E: GenericEvent>(&mut self, event: &E)

Handle game events and update the state.

fn pick_canvas(&self, xy: Point) -> Option<CanvasId>

If the given Point is currently on a Canvas, return the Id of that canvas.

fn get_xy(&self, position: Position, dim: Dimensions, h_align: HorizontalAlign, v_align: VerticalAlign) -> Point

Get the centred xy coords for some given Dimensions, Position and alignment.

fn draw<G>(&mut self, graphics: &mut G) where C: CharacterCache, G: Graphics<Texture=C>

Draw the Ui in it's current state. - The order of drawing is as follows: 1. Canvas splits. 2. Widgets on Canvas splits. 3. Floating Canvasses. 4. Widgets on Floating Canvasses. - Widgets are sorted by capturing and then render depth (depth first). - Construct the elmesque Renderer for rendering the elm Elements. - Render all widgets.