Struct conrod::UiCell

pub struct UiCell<'a> { /* fields omitted */ }

A wrapper around the Ui that restricts the user from mutating the Ui in certain ways while in the scope of the Ui::set_widgets function and within Widgets' update methods. Using the UiCell, users may access the Ui immutably (via Deref) however they wish, however they may only mutate the Ui via the &mut self methods provided by the UiCell.

The name came from its likening to a "jail cell for the Ui", as it restricts a user's access to it. However, we realise that the name may also cause ambiguity with the std Cell and RefCell types (which UiCell has nothing to do with). Thus, if you have a better name for this type in mind, please let us know at the github repo via an issue or PR sometime before we hit 1.0.0!

Methods

impl<'a> UiCell<'a>

fn theme(&self) -> &Theme

A reference to the Theme that is currently active within the Ui.

fn font(&self, id: Id) -> Option<&Font>

A convenience method for borrowing the Font for the given Id if it exists.

fn window_dim(&self) -> Dimensions

Returns the dimensions of the window

fn global_input(&self) -> &Global

Returns an immutable reference to the input::Global of the Ui.

All coordinates here will be relative to the center of the window.

fn widget_input(&self, id: Id) -> Widget

Returns a input::Widget with input events for the widget.

All coordinates in the input::Widget will be relative to the widget at the given index.

fn widget_id_generator(&mut self) -> Generator

Produces a type that may be used to generate new unique widget::Ids.

See the widget::id::Generator docs for details on how to use this correctly.

fn kids_bounding_box(&self, id: Id) -> Option<Rect>

The Rect that bounds the kids of the widget with the given index.

Returns None if the widget has no children or if there's is no widget for the given index.

fn scroll_widget(&mut self, id: Id, offset: [Scalar; 2])

Scroll the widget at the given index by the given offset amount.

The produced Scroll event will be pushed to the pending_scroll_events and will be applied to the widget during the next call to Ui::set_widgets.