Struct conrod::UiCell

pub struct UiCell<'a, B: 'a> where B: Backend {
    // some fields omitted
}

A wrapper around the Ui that restricts the user from mutating the Ui in certain ways while in the scope of the Ui::set_widgets function and within Widgets' update methods. Using the UiCell, users may access the Ui immutably (via Deref) however they wish, however they may only mutate the Ui via the &mut self methods provided by the UiCell.

The name came from its likening to a "jail cell for the Ui", as it restricts a user's access to it. However, we realise that the name may also cause ambiguity with the std Cell and RefCell types (which UiCell has nothing to do with). Thus, if you have a better name for this type in mind, please let us know at the github repo via an issue or PR sometime before we hit 1.0.0!

Methods

impl<'a, B> UiCell<'a, B> where B: Backend

fn theme(&self) -> &Theme

A reference to the Theme that is currently active within the Ui.

fn glyph_cache(&self) -> &GlyphCache<B::CharacterCache>

A reference to the Ui's GlyphCache.

fn window_dim(&self) -> Dimensions

Returns the dimensions of the window

fn global_input(&self) -> &Global

Returns an immutable reference to the input::Global of the Ui.

All coordinates here will be relative to the center of the window.

fn widget_input<I: Into<Index>>(&self, idx: I) -> Widget

Returns a input::Widget with input events for the widget.

All coordinates in the input::Widget will be relative to the widget at the given index.

fn new_unique_node_index(&mut self) -> NodeIndex

Generate a new, unique NodeIndex into a Placeholder node within the Ui's widget graph. This should only be called once for each unique widget needed to avoid unnecessary bloat within the Ui's widget graph.

When using this method in your Widget's update method, be sure to store the returned NodeIndex somewhere within your Widget::State so that it can be re-used on next update.

Panics if adding another node would exceed the maximum capacity for node indices.

fn kids_bounding_box<I: Into<Index>>(&self, idx: I) -> Option<Rect>

The Rect that bounds the kids of the widget with the given index.

Returns None if the widget has no children or if there's is no widget for the given index.

fn scroll_widget<I>(&mut self, widget_idx: I, offset: [Scalar; 2]) where I: Into<Index>

Scroll the widget at the given index by the given offset amount.

The produced Scroll event will be pushed to the pending_scroll_events and will be applied to the widget during the next call to Ui::set_widgets.