Struct conrod::UpdateArgs

pub struct UpdateArgs<'a, 'b, W, C: 'a> where W: Widget {
    pub prev_state: &'a State<W::State>,
    pub xy: Point,
    pub dim: Dimensions,
    pub input: UserInput<'b>,
    pub style: &'a W::Style,
    pub theme: &'a Theme,
    pub glyph_cache: &'a GlyphCache<C>,
}

Arguments for the Widget::update method in a struct to simplify the method signature.

Fields

prev_state

The Widget's state that was last returned by the update method.

xy

The absolute (centered origin) screen position of the widget.

dim

The dimensions of the Widget.

input

The current state of user input (i.e. mouse, keys pressed, etc).

style

The Widget's current Style.

theme

The active Theme within the Ui.

glyph_cache

The Ui's GlyphCache (for determining text width, etc).