Struct conrod::UpdateArgs

pub struct UpdateArgs<'a, 'b, W, C: 'a> where W: Widget {
    pub idx: Index,
    pub prev: &'a CommonState,
    pub state: &'a mut State<'b, W::State>,
    pub rect: Rect,
    pub style: &'a W::Style,
    pub ui: UiCell<'a, C>,
}

Arguments for the Widget::update method in a struct to simplify the method signature.

Fields

idx

The Widget's unique index.

prev

The Widget's previous state. Specifically, the state that is common between all widgets, such as positioning, floatability, draggability, etc.

state

A wrapper around the Widget's unique state, providing methods for both immutably viewing and mutably updating the state.

We wrap mutation in a method so that we can keep track of whether or not the unique state has been updated.

If State::update is called, we assume that there has been some mutation and in turn will produce a new Element for the Widget. Thus, it is recommended that you only call State::update if you need to update the unique state in some way.

rect

The rectangle describing the Widget's area.

style

The Widget's current Widget::Style.

ui

Restricted access to the Ui.

Provides methods for immutably accessing the Ui's Theme and GlyphCache. Also allows calling Widget::set within the Widget::update method.