Struct conrod::WidgetMatrix

pub struct WidgetMatrix<F> {
    // some fields omitted
}

Draw a matrix of any rectangular widget type, where the matrix will provide a function with the widget number, it's rows and cols position, the width and height for the widget and the location at which the widget should be drawn.

Methods

impl<F> Matrix<F>

fn new(cols: usize, rows: usize) -> Matrix<F>

Create a widget matrix context.

fn each_widget(self, each_widget: F) -> Matrix<F>

The function that will be called for each and every element in the Matrix. The function should return the widget that will be displayed in the element associated with the given row and column number. Note that the returned Widget's position and dimensions will be overridden with the dimensions and position of the matrix element's rectangle.

fn cell_padding(self, w: Scalar, h: Scalar) -> Matrix<F>

A builder method for adding padding to the cell.