Struct conrod::WidgetMatrix

pub struct WidgetMatrix {
    // some fields omitted
}

Draw a matrix of any rectangular widget type, where the matrix will provide a function with the widget number, it's rows and cols position, the width and height for the widget and the location at which the widget should be drawn.

Methods

impl Matrix

fn new(cols: usize, rows: usize) -> Matrix

Create a widget matrix context.

fn each_widget<C, F>(&mut self, ui: &mut Ui<C>, react: F) where F: FnMut(&mut Ui<C>, WidgetNum, ColNum, RowNum, Point, Dimensions)

The reaction called for each widget in the matrix. This should be called following all builder methods.

fn cell_padding(self, w: f64, h: f64) -> Matrix

A builder method for adding padding to the cell.