Struct conrod::WidgetState

pub struct WidgetState<T> {
    pub state: T,
    pub dim: Dimensions,
    pub xy: Point,
    pub depth: Depth,
    pub drag_state: State,
    pub maybe_floating: Option<Floating>,
}

Represents the unique cached state of a widget.

Fields

state

The state of the Widget.

dim

The rectangular dimensions of the Widget.

xy

The position of the Widget given as [x, y] coordinates.

depth

The rendering depth for the Widget (the default is 0.0).

drag_state

The current state of the dragged widget, if it is draggable.

maybe_floating

Floating state for the widget if it is floating.