Struct conrod::XYPad

pub struct XYPad<'a, X, Y, F> {
    pub maybe_react: Option<F>,
    pub enabled: bool,
    // some fields omitted
}

Used for displaying and controlling a 2D point on a cartesian plane within a given range.

Its reaction is triggered when the value is updated or if the mouse button is released while the cursor is above the rectangle.

Fields
maybe_react: Option<F>

The reaction function for the XYPad.

It will be triggered when the value is updated or if the mouse button is released while the cursor is above the rectangle.

enabled: bool

Indicates whether the XYPad will respond to user input.

Methods

impl<'a, X, Y, F> XYPad<'a, X, Y, F>

fn new(x_val: X, min_x: X, max_x: X, y_val: Y, min_y: Y, max_y: Y) -> Self

Build a new XYPad widget.

fn line_thickness(self, line_thickness: Scalar) -> Self

Build the type's self.$($assignee).+ with the given $Type.

fn value_font_size(self, value_font_size: FontSize) -> Self

Build the type's self.$($assignee).+ with the given $Type.

fn react(self, react: F) -> Self

Build the type's self.$($assignee).+ with the given $Type.

fn enabled(self, enabled: bool) -> Self

Build the type's self.$($assignee).+ with the given $Type.