Struct conrod::theme::Theme

pub struct Theme {
    pub name: String,
    pub padding: Padding,
    pub x_position: Position,
    pub y_position: Position,
    pub background_color: Color,
    pub shape_color: Color,
    pub border_color: Color,
    pub border_width: Scalar,
    pub label_color: Color,
    pub font_id: Option<Id>,
    pub font_size_large: u32,
    pub font_size_medium: u32,
    pub font_size_small: u32,
    pub widget_styling: HashMap<TypeId, WidgetDefault>,
    pub mouse_drag_threshold: Scalar,
    pub double_click_threshold: Duration,
}

A serializable collection of canvas and widget styling defaults.

Fields
name: String

A name for the theme used for identification.

padding: Padding

Padding for Canvas layout and positioning.

x_position: Position

A default widget position along the x axis.

y_position: Position

A default widget position along the y axis.

background_color: Color

A default background for the theme.

shape_color: Color

A default color for widget shapes.

border_color: Color

A default color for widget borders.

border_width: Scalar

A default width for widget borders.

label_color: Color

A default color for widget labels.

font_id: Option<Id>

The Id of the default font used for text widgets when one is not specified.

font_size_large: u32

A default "large" font size.

font_size_medium: u32

A default "medium" font size.

font_size_small: u32

A default "small" font size.

widget_styling: HashMap<TypeId, WidgetDefault>

Unique styling for each widget, index-able by the Widget::kind.

mouse_drag_threshold: Scalar

Mouse Drag distance threshold determines the minimum distance from the mouse-down point that the mouse must move before starting a drag operation.

double_click_threshold: Duration

Once the Duration that separates two consecutive Clicks is greater than this value, a DoubleClick event will no longer be generated.

Methods

impl Theme

fn default() -> Theme

The default theme if not loading from file.

fn widget_style<T>(&self) -> Option<UniqueDefault<T>> where T: Style

Retrieve the unique default styling for a widget.

Attempts to cast the Box<WidgetStyle> to the Widget's unique associated style T.