Trait conrod::Widget

pub trait Widget: Positionable + Sizeable + Sized {
    type State: Any + PartialEq + Debug;
    type Style: Any + PartialEq + Debug;
    fn unique_kind(&self) -> &'static str;
    fn init_state(&self) -> Self::State;
    fn style(&self) -> Self::Style;
    fn update<'a, C>(self, prev: &State<Self::State>, xy: Point, dim: Dimensions, input: UserInput<'a>, current_style: &Self::Style, theme: &Theme, glyph_cache: &GlyphCache<C>) -> Option<Self::State> where C: CharacterCache;
    fn draw<C>(new_state: &State<Self::State>, current_style: &Self::Style, theme: &Theme, glyph_cache: &GlyphCache<C>) -> Element where C: CharacterCache;

    fn set<C>(self, id: WidgetId, ui: &mut Ui<C>) where C: CharacterCache { ... }
    fn capture_mouse(_prev: &Self::State, _new: &Self::State) -> bool { ... }
    fn uncapture_mouse(_prev: &Self::State, _new: &Self::State) -> bool { ... }
    fn capture_keyboard(_prev: &Self::State, _new: &Self::State) -> bool { ... }
    fn uncapture_keyboard(_prev: &Self::State, _new: &Self::State) -> bool { ... }
    fn canvas_id(&self) -> Option<CanvasId> { ... }
}

A trait to be implemented by all Widget types.

Associated Types

type State: Any + PartialEq + Debug

State to be stored within the Uis widget cache. Take advantage of this type for any large allocations that you would like to avoid repeating between updates, or any calculations that you'd like to avoid repeating between calls to update and draw. Conrod will never clone the state, it will only ever be moved.

type Style: Any + PartialEq + Debug

Styling used by the widget to construct an Element. Styling is useful to have in its own abstraction in order to making Theme serializing easier. Conrod doesn't yet support serializing non-internal widget styling with the Theme type, but we hope to soon.

Required Methods

fn unique_kind(&self) -> &'static str

Return the kind of the widget as a &'static str. Note that this must be unique from all other widgets' "unique kinds". This is used by conrod to help avoid UiId errors.

fn init_state(&self) -> Self::State

Return the initial State of the Widget. The Ui will only call this once.

fn style(&self) -> Self::Style

Return the styling of the widget. The Ui will call this once prior to each update. It does this so that it can check for differences in Style in case a new Element needs to be constructed.

fn update<'a, C>(self, prev: &State<Self::State>, xy: Point, dim: Dimensions, input: UserInput<'a>, current_style: &Self::Style, theme: &Theme, glyph_cache: &GlyphCache<C>) -> Option<Self::State> where C: CharacterCache

Your widget's previous state is given to you as a parameter and it is your job to construct and return an Update that will be used to update the widget's cached state.

fn draw<C>(new_state: &State<Self::State>, current_style: &Self::Style, theme: &Theme, glyph_cache: &GlyphCache<C>) -> Element where C: CharacterCache

Construct a renderable Element from the current styling and new state. This will only be called on the occasion that the widget's Style or State has changed. Keep this in mind when designing your widget's Style and State types.

Provided Methods

fn set<C>(self, id: WidgetId, ui: &mut Ui<C>) where C: CharacterCache

After building the widget, you call this method to set its current state into the given Ui. More precisely, the following will occur when calling this method: - The widget's previous state and style will be retrieved. - The widget's current Style will be retrieved (from the Widget::style method). - The widget's state will be updated (using the Widget::udpate method). - If the widget's state or style has changed, Widget::draw will be called to create the new Element for rendering. - The new State, Style and Element (if there is one) will be cached within the Ui.

fn capture_mouse(_prev: &Self::State, _new: &Self::State) -> bool

Optionally override with the case that the widget should capture the mouse.

fn uncapture_mouse(_prev: &Self::State, _new: &Self::State) -> bool

Optionally override with the case that the widget should capture the mouse.

fn capture_keyboard(_prev: &Self::State, _new: &Self::State) -> bool

Optionally override with the case that the widget should capture the mouse.

fn uncapture_keyboard(_prev: &Self::State, _new: &Self::State) -> bool

Optionally override with the case that the widget should capture the mouse.

fn canvas_id(&self) -> Option<CanvasId>

Return the canvas to which the Widget will be attached, if there is one.