Module conrod::utils

Various utility functions used throughout Conrod.

Enums

IterDiff
A type returned by the iter_diff function.

Functions

bounding_box_for_points
Find the bounding rect for the given series of points.

clamp
Clamp a value between some range.

degrees
Convert degrees to radians.

fmod
Modulo float.

iter_diff
Compares every element yielded by both elems and new_elems in lock-step.

iter_eq
Compares two iterators to see if they yield the same thing.

map_range
Map a value from a given range to a new given range.

modulo
The modulo function.

partial_max
Compare to PartialOrd values and return the min.

partial_min
Compare to PartialOrd values and return the min.

percentage
Get value percentage between max and min.

turns
Convert turns to radians.

val_to_string
Get a suitable string from the value, its max and the pixel range.

value_from_perc
Adjust the value to the given percentage.

vec2_add
Add a and b.

vec2_sub
Subtract b from a.

write_if_different
Returns Borrowed elems if elems contains the same elements as yielded by new_elems.