Struct conrod::widget::bordered_rectangle::BorderedRectangle

pub struct BorderedRectangle {
    pub common: CommonBuilder,
    pub style: Style,
}

A filled rectangle widget that may or may not have some border.

NOTE: BorderedRectangle is currently implemented as two filled rectangles:

A Rectangle for the border.

A Rectangle for the non-border area.

This is flawed in that, if a user specifies an alpha lower than 1.0, the front Rectangle will blend with the border Rectangle, which is likely unexpected behaviour. This should be changed so that the border is drawn using a outlined Rectangle.

Fields
common: CommonBuilder

Data necessary and common for all widget builder types.

style: Style

Unique styling for the BorderedRectangle.

Methods

impl BorderedRectangle

pub fn new(dim: Dimensions) -> Self

Build a new BorderedRectangle.

pub fn with_style(self, with_style: Style) -> Self

Build the type's self.$($assignee).+ with the given $Type.