Struct conrod::widget::button::Button

pub struct Button<'a, S> {
    pub show: S,
    pub style: Style,
    // some fields omitted
}

A pressable button widget whose reaction is triggered upon release.

Fields
show: S

Whether the Button is a Flat color or an Image.

style: Style

Unique styling parameters for the Button.

Methods

impl<'a> Button<'a, Image>

fn image(image_id: Id) -> Self

Begin building a button displaying the given Image on top.

fn source_rectangle(self, rect: Rect) -> Self

The rectangular area of the image that we wish to display.

If this method is not called, the entire image will be used.

fn image_color(self, color: Color) -> Self

Map the Image's luminance to the given color.

fn image_color_with_feedback(self, color: Color) -> Self

Map the Image's luminance to the given color.

The color will change slightly when the button is highlighted or clicked to give the user some visual feedback.

fn hover_image(self, id: Id) -> Self

The image displayed while the mouse hovers over the Button.

fn press_image(self, id: Id) -> Self

The image displayed while the Button is pressed.

impl<'a> Button<'a, Flat>

fn new() -> Self

Begin building a flat-colored Button widget.

impl<'a, S> Button<'a, S>

fn label_font_id(self, font_id: Id) -> Self

Specify the font used for displaying the label.

fn left_justify_label(self) -> Self

Align the label to the left of the Button's surface.

fn center_justify_label(self) -> Self

Align the label to the mid-left of the Button's surface.

This is the default label alignment.

fn right_justify_label(self) -> Self

Align the label to the mid-left of the Button's surface.

fn label_x(self, x: Relative) -> Self

Specify the label's position relatively to Button along the x axis.

fn label_y(self, y: Relative) -> Self

Specify the label's position relatively to Button along the y axis.

fn enabled(self, enabled: bool) -> Self

Build the type's self.$($assignee).+ with the given $Type.