Struct conrod::widget::button::Image

pub struct Image {
    pub image_id: Id,
    pub hover_image_id: Option<Id>,
    pub press_image_id: Option<Id>,
    pub color: ImageColor,
    pub src_rect: Option<Rect>,
}

The Button displays an Image on top.

Fields
image_id: Id

The id of the Image to be used.

hover_image_id: Option<Id>

The image displayed when the mouse hovers over the button.

press_image_id: Option<Id>

The image displayed when the mouse has captured and is pressing the button.

color: ImageColor

If Some, maps the image's luminance to this Color.

src_rect: Option<Rect>

The rectangular area of the original source image that should be displayed.