Enum conrod::widget::MaybeParent

pub enum MaybeParent {
    None,
    Some(Id),
    Unspecified,
}

The builder argument for the Widget's parent.

Variants
None

The user specified the widget should not have any parents, so the Root will be used.

Some(Id)

The user gave a specific parent widget.

Unspecified

No parent widget was specified, so we will assume they want the last parent.

Methods

impl MaybeParent

fn get_unchecked(&self, ui: &Ui, x_pos: Position, y_pos: Position) -> Id

Convert the MaybeParent into an Option.

If Unspecified, check the positioning to retrieve the Id from there.

If None, the Ui's window widget will be used.

Note: This method does not check whether or not using the window widget as the parent would cause a cycle. If it is important that the inferred parent should not cause a cycle, use get instead.

fn get(&self, id: Id, ui: &Ui, x_pos: Position, y_pos: Position) -> Option<Id>

The same as get_unchecked, but checks whether or not the widget that we're inferring the parent for is the Ui's window (which cannot have a parent, without creating a cycle).