Struct conrod::widget::grid::Grid

pub struct Grid<X, Y, I> {
    pub common: CommonBuilder,
    pub style: Style,
    pub min_x: X,
    pub max_x: X,
    pub min_y: Y,
    pub max_y: Y,
    pub x_offset: Option<X>,
    pub y_offset: Option<Y>,
    pub lines: I,
}

A widget for displaying a grid of lines across two axes.

Fields
common: CommonBuilder

Builder parameters that are common to all Widgets.

style: Style

Unique styling parameters for the Grid widget.

min_x: X

The minimum visible bound along the x axis.

max_x: X

The maximum visible bound along the x axis.

min_y: Y

The minimum visible bound along the y axis.

max_y: Y

The maximum visible bound along the y axis.

x_offset: Option<X>

An offset for all vertical lines distributed across the x axis.

y_offset: Option<Y>

An offset for all horizontal lines distributed across the y axis.

lines: I

An iterator yielding each sequence of lines to be distributed across the grid.

Methods

impl<X, Y, I> Grid<X, Y, I>

pub fn new(
min_x: X,
max_x: X,
min_y: Y,
max_y: Y,
lines: I
) -> Grid<X, Y, I::IntoIter> where
X: Into<Scalar>,
Y: Into<Scalar>,
I: IntoIterator<Item = Axis<X, Y>>,

Begin building a new PlotPath widget instance.

The first four arguments represent the visible range along both axes.

The final argument is an iterator yielding Lines across either Axis. The given lines will be drawn in the order that they're given.

pub fn x_offset(self, x: X) -> Self

Specify an offset for all vertical lines placed along the X axis.

pub fn y_offset(self, y: Y) -> Self

Specify an offset for all horizontal lines placed along the Y axis.