Struct conrod::widget::list::Item

pub struct Item<D, S> {
    pub i: usize,
    pub widget_id: Id,
    pub last_id: Option<Id>,
    // some fields omitted
}

The data necessary for instantiating a single item within a List.

Fields
i: usize

The index of the item within the list.

widget_id: Id

The id generated for the widget.

last_id: Option<Id>

The id used for the previous item's widget.

Methods

impl<D, S> Item<D, S> where
D: Direction,
S: ItemSize,

pub fn set<W>(self, widget: W, ui: &mut UiCell) -> W::Event where
W: Widget,

Sets the given widget as the widget to use for the item.

Sets the:

position of the widget.

dimensions of the widget.

parent of the widget.

and finally sets the widget within the Ui.

impl<S> Item<Down, S>

pub fn width(&self) -> Scalar

The width of the Item.

impl Item<Down, Fixed>

pub fn height(&self) -> Scalar

The height of the Item.

impl<S> Item<Up, S>

pub fn width(&self) -> Scalar

The width of the Item.

impl Item<Up, Fixed>

pub fn height(&self) -> Scalar

The height of the Item.

impl<S> Item<Right, S>

pub fn height(&self) -> Scalar

The height of the Item.

impl Item<Right, Fixed>

pub fn width(&self) -> Scalar

The width of the Item.

impl<S> Item<Left, S>

pub fn height(&self) -> Scalar

The height of the Item.

impl Item<Left, Fixed>

pub fn width(&self) -> Scalar

The width of the Item.