Struct conrod::widget::list::List

pub struct List {
    pub common: CommonBuilder,
    pub style: Style,
    pub item_instantiation: ItemInstantiation,
    // some fields omitted
}

A helper widget, useful for instantiating a sequence of widgets in a vertical list.

The List widget simplifies this process by:

Generating widget::Ids.

Simplifying the positioning and sizing of items.

Optimised widget instantiation by only instantiating visible items. This is very useful for lists containing many items, i.e. a FileNavigator over a directory with thousands of files.

Fields
common: CommonBuilder

Common widget building params for the List.

style: Style

Unique styling for the List.

item_instantiation: ItemInstantiation

Whether all or only visible items should be instantiated.

Methods

impl List

fn new(num_items: usize, item_height: Scalar) -> Self

Create a List context to be built upon.

fn scrollbar_next_to(self) -> Self

Specifies that the List should be scrollable and should provide a Scrollbar to the right of the items.

fn scrollbar_on_top(self) -> Self

Specifies that the List should be scrollable and should provide a Scrollbar that hovers above the right edge of the items and automatically hides when the user is not scrolling.

fn scrollbar_width(self, w: Scalar) -> Self

The width of the Scrollbar.

fn scrollbar_color(self, color: Color) -> Self

The color of the Scrollbar.

fn instantiate_all_items(self) -> Self

Indicates that an Item should be instatiated for every element in the list, regardless of whether or not the Item would be visible.

Note: This may cause significantly heavier CPU load for lists containing many items (100+). We only recommend using this when absolutely necessary as large lists may cause unnecessary bloating within the widget graph, and in turn result in greater traversal times.

fn instantiate_only_visible_items(self) -> Self

Indicates that only Items that are visible should be instantiated. This ensures that we avoid bloating the widget graph with unnecessary nodes and in turn keep traversal times to a minimum.

This is the default List behaviour.