Struct conrod::widget::tabs::TabRects

pub struct TabRects<'a> { /* fields omitted */ }

An iterator yielding the Rect for each Tab in the given list.

Methods

impl<'a> TabRects<'a>

fn new(tabs: &'a [(Id, &'a str)],
layout: Layout,
rel_tab_bar_rect: Rect)
-> Self

Construct a new TabRects iterator.

fn next_with_id_and_label(&mut self) -> Option<(Rect, Id, &'a str)>

Yield the next Tab Rect, along with the associated ID and label.