Struct conrod::widget::text_edit::TextEdit

pub struct TextEdit<'a> { /* fields omitted */ }

A widget for displaying and mutating multi-line text, given as a String.

By default the text is wrapped via the first whitespace before the line exceeds the TextEdit's width, however a user may change this using the .wrap_by_character method.

Methods

impl<'a> TextEdit<'a>

fn new(text: &'a str) -> Self

Construct a TextEdit widget.

fn wrap_by_whitespace(self) -> Self

The TextEdit will wrap text via the whitespace that precedes the first width-exceeding character.

This is the default setting.

fn wrap_by_character(self) -> Self

By default, the TextEdit will wrap text via the whitespace that precedes the first width-exceeding character.

Calling this method causes the TextEdit to wrap text at the first exceeding character.

fn left_justify(self) -> Self

Align the text to the left of its bounding Rect's x axis range.

fn center_justify(self) -> Self

Align the text to the middle of its bounding Rect's x axis range.

fn right_justify(self) -> Self

Align the text to the right of its bounding Rect's x axis range.

fn align_text_bottom(self) -> Self

Align the text to the left of its bounding Rect's y axis range.

fn align_text_y_middle(self) -> Self

Align the text to the middle of its bounding Rect's y axis range.

fn align_text_top(self) -> Self

Align the text to the right of its bounding Rect's y axis range.

fn align_text_middle(self) -> Self

Align the text to the middle of its bounding Rect.

fn font_id(self, font_id: Id) -> Self

Specify the font used for displaying the text.

fn font_size(self, font_size: FontSize) -> Self

Build the type's self.$($assignee).+ with the given $Type.

fn justify(self, justify: Justify) -> Self

Build the type's self.$($assignee).+ with the given $Type.

fn y_align_text(self, y_align_text: Align) -> Self

Build the type's self.$($assignee).+ with the given $Type.

fn line_wrap(self, line_wrap: Wrap) -> Self

Build the type's self.$($assignee).+ with the given $Type.

fn line_spacing(self, line_spacing: Scalar) -> Self

Build the type's self.$($assignee).+ with the given $Type.

fn restrict_to_height(self, restrict_to_height: bool) -> Self

Build the type's self.$($assignee).+ with the given $Type.